Module nx::arm

ARM support and utils

Structs

CpuRegister

Represents a CPU register value (W, X or R value depending on the context/arch)

FpuRegister

Represents a FPU register value (V, D or S value depending on the context/arch)

RegisterGroup

Represents flags of different register kinds/groups

ThreadContext

Represents a thread context usable with svcs

Functions

cache_flush

Flushes memory cache

get_system_tick

Gets the system tick

get_system_tick_frequency

Gets the system tick frequency

nanoseconds_to_ticks

Converts nanoseconds to ticks

ticks_to_nanoseconds

Converts ticks to nanoseconds