nx::crypto::aes::a128
Struct Context

pub struct Context {
    pub round_keys: [[u8; 16]; 11],
}

Represents the context used for 128-bit AES operations

Fields
round_keys: [[u8; 16]; 11]

The round keys

Implementations

impl Context

pub fn new(key: &[u8], is_encryptor: bool) -> Result<Self>

Creates a new Context with the given key

The key must have size KEY_SIZE in bytes or this will fail with ResultInvalidSize

Arguments

key: The key to use

is_encryptor: Whether this context will be used for encrypting or decrypting

pub fn encrypt_block(&self, src: &[u8], dst: &mut [u8])

Encrypts the given data

Arguments

src: The input data

dst: The output data to fill into

pub fn decrypt_block(&self, src: &[u8], dst: &mut [u8])

Decrypts the given data

Arguments

src: The input data

dst: The output data to fill into