Struct nx::crypto::aes::ctr::a128::Context

pub struct Context { /* private fields */ }

Represents the context used for 128-bit AES-CTR operations

Implementations

impl Context

pub fn new(key: &[u8], ctr: &[u8]) -> Result<Self>

Creates a new Context

The key must have size KEY_SIZE in bytes and the ctr must have size BLOCK_SIZE in bytes or this will fail with ResultInvalidSize

Arguments

key: The 128-bit AES key

ctr: The ctr

pub fn reset_ctr(&mut self, ctr: &[u8]) -> Result<()>

Resets the Context with the provided ctr

The ctr must have size BLOCK_SIZE in bytes or this will fail with ResultInvalidSize

Argument

ctr: The ctr to reset with

pub fn crypt(&mut self, src: &[u8], dst: &mut [u8]) -> Result<()>

Crypts the given data

Input and output data must have the same size, or this will fail with ResultInvalidSize

Arguments

src: The input data

dst: The output data to fill into