nx::crypto::hmac::sha256

Function calculate_mac

pub fn calculate_mac<T>(
    key: &[u8],
    data: &[u8],
    out_mac: &mut [T],
) -> Result<()>

Wrapper for directly calculating the MAC of given data

The output hash array must have size HASH_SIZE in bytes or this will fail with ResultInvalidSize

This essentially creates a Context, updates it with the given data and produces its MAC

Arguments

key: Input key

data: Input data

out_,ac: Output array to fill into