nx::elf::mod0

Struct Header

#[repr(C)]
pub struct Header { /* private fields */ }

Represents the MOD0 header structure. Although we know from the official linker script that all the offsets will be positive, the offsets have been made signed so that consumers can bring their own linker scripts (e.g. emuiibo) and we wonโt break functionality if they reorder sections.

All members have been made private as this should only ever be instantiated using Header::from_text_start_addr, with data populated by the linker.

Implementations

impl Header

pub const MAGIC: u32 = 809_783_117u32

The header magic value (MOD0)

pub fn from_text_start_addr(text_base: *mut u8) -> &'static mut Self

Gets the header embedded at the slot .text.jmp+4. Since this is a hard requirement of the switch runtime, an invalid MOD0 header offset or invalid header magic value panics. Panics if the text_base pointer is invalid, the derived MOD0 pointer is invalid, or if the derived MOD0 header magic value is incorrect.

Arguments:

text_base: The start of the .text.jmp section.

pub fn get_dyn_start(&self) -> *const Dyn

Gets the start address for the .dynamic section

pub fn get_eh_frame_header_start(&self) -> *const u8

Gets the start address for the eh_frame_hdr section.

Safety

The reference &self must be the copy in .rodata created by the linker.

pub unsafe fn zero_bss_section(&mut self)

Zeroes the bss section from a base code address. We have to take an &mut self here as computing.

Safety

The reference &mut self must be the copy in .rodata created by the linker. Additionally, The reference to self should have been created using a mutable pointer, to prevent a shared->mut conversion which would be immediate UB (as documented in the struct docstring).