Function nx::fs::format_path

pub fn format_path(path: String) -> Result<(Shared<dyn FileSystem>, String)>

Returns the FileSystem and the processed path corresponding to a given path

If there is a filesystem mounted as demo, calling this with "demo:/anything" will return an instance to that mounted filesystem and "anything" as the processed path

Arguments

path: The path to use