Trait nx::fs::Directory

pub trait Directory {
    // Required methods
    fn read(&mut self, out_entries: &mut [DirectoryEntry]) -> Result<u64>;
    fn get_entry_count(&mut self) -> Result<u64>;
}

Represents a directory

Required Methods

fn read(&mut self, out_entries: &mut [DirectoryEntry]) -> Result<u64>

Reads existing entries, returning the actual number of read entries

The max number of entries to read is determined by the output array size and the actually existing entry count

Arguments

out_entries: The out DirectoryEntry array to fill

fn get_entry_count(&mut self) -> Result<u64>

Gets the Directory’s entry count