nx::fs

Trait File

pub trait File: Sync {
    // Required methods
    fn read(
        &mut self,
        offset: usize,
        out_buf: &mut [u8],
        option: FileReadOption,
    ) -> Result<usize>;
    fn write(
        &mut self,
        offset: usize,
        buf: &[u8],
        option: FileWriteOption,
    ) -> Result<()>;
    fn flush(&mut self) -> Result<()>;
    fn set_size(&mut self, size: usize) -> Result<()>;
    fn get_size(&mut self) -> Result<usize>;
    fn operate_range(
        &mut self,
        operation_id: OperationId,
        offset: usize,
        size: usize,
    ) -> Result<FileQueryRangeInfo>;
    fn operate_range_with_buffer(
        &mut self,
        operation_id: OperationId,
        offset: usize,
        size: usize,
        in_buf: &[u8],
        out_buf: &mut [u8],
    ) -> Result<()>;
}

Represents a file, abstracted from the IPC client API.

Required Methods

fn read( &mut self, offset: usize, out_buf: &mut [u8], option: FileReadOption, ) -> Result<usize>

Reads data from the file, returning the actual read size.

Arguments:

offset: The absolute offset.

out_buf: The output slice to fill.

option: FileReadOption for file reading flags.

fn write( &mut self, offset: usize, buf: &[u8], option: FileWriteOption, ) -> Result<()>

Writes data to a file (this one doesn’t return the actual written size, thanks N).

Arguments.

offset: The absolute offset.

buf: The input data to write into the file.

option: FileWriteOption value.

fn flush(&mut self) -> Result<()>

Flushes the pending file writes.

fn set_size(&mut self, size: usize) -> Result<()>

Sets the file size.

This effectively truncates the file.

Arguments.

size: The new file size.

fn get_size(&mut self) -> Result<usize>

Gets the current file size.

fn operate_range( &mut self, operation_id: OperationId, offset: usize, size: usize, ) -> Result<FileQueryRangeInfo>

Performs a range-operation on the file, returning corresponding result data.

Arguments:

operation_id: The ID of the file operation to perform on the specified range.

offset: The absolute offset.

size: The file data size in which to operate. i.e. we are operating in the range [offset, offset+size).

fn operate_range_with_buffer( &mut self, operation_id: OperationId, offset: usize, size: usize, in_buf: &[u8], out_buf: &mut [u8], ) -> Result<()>

Performs a range-operation on the file with custom input/output data.

Arguments:

operation_id: The ID of the file operation to perform on the specified range.

offset: The absolute offset.

size: The file data size in which to operate. i.e. we are operating in the range [offset, offset+size).

in_buf: Input data buffer.

out_buf: Output data buffer.