nx::gpu::canvas

Struct CanvasManager

pub struct CanvasManager<ColorFormat: CanvasColorFormat> {
    pub surface: Surface,
    /* private fields */
}

Fields
surface: Surface
Implementations

impl<ColorFormat: CanvasColorFormat> CanvasManager<ColorFormat>

pub const fn total_heap_required( width: u32, height: u32, block_height_config: BlockLinearHeights, buffer_count: u32, ) -> usize

pub fn new_managed( gpu_ctx: Arc<RwLock<Context>>, surface_name: Option<&'static str>, x: u32, y: u32, z: LayerZ, width: u32, height: u32, aruid: AppletResourceUserId, layer_flags: LayerFlags, buffer_count: u32, block_height: BlockLinearHeights, scaling: ScaleMode, ) -> Result<Self>

Creates a new stray layer (application/applet window) that can be drawn on for UI elements

pub fn new_stray( gpu_ctx: Arc<RwLock<Context>>, surface_name: Option<&'static str>, buffer_count: u32, block_height: BlockLinearHeights, ) -> Result<Self>

Creates a new managed layer (application/applet window) that can be drawn on overlay elements. These layers exist on top of stray layers and application/applet UIs, and can be Z-order vertically layered over each other. The GPU provides the alpha blending for all layers based on the committed frame.

pub fn wait_vsync_event(&self, timeout: Option<i64>) -> Result<()>

Wait for a vsync event to ensure that the previously submitted frame has been fully rendered to the display.

pub fn render<T>( &mut self, clear_color: Option<ColorFormat>, runner: impl Fn(&mut BufferedCanvas<'_, ColorFormat>) -> Result<T>, ) -> Result<T>

Check out a canvas/framebuffer to draw a new frame. This frame is buffered to a linear buffer before being swizzled into the backing buffer during frame-commit. This provides better performance for line-based renderers as writes will be sequential in memory. There are cache misses during commit, but that is still better performance than mapped page churn for GPU-mapped memory.

pub fn render_unbuffered<T>( &mut self, clear_color: Option<ColorFormat>, runner: impl Fn(&mut UnbufferedCanvas<'_, ColorFormat>) -> Result<T>, ) -> Result<T>

Check out a canvas/framebuffer to draw a new frame, without a linear buffer. This can be used in memory constrained environments such as sysmodules, but also provides slightly better performance for frames that draw in block rather than scan-lines (e.g. JPEG decoding).