nx::gpu::parcel

Struct Parcel

pub struct Parcel { /* private fields */ }

Represents a wrapper for simple parcel reading/writing

Implementations

impl Parcel

pub const fn new() -> Self

Creates a new Parcel

pub unsafe fn read_raw_unaligned( &mut self, out_data: *mut u8, data_size: usize, ) -> Result<()>

Reads raw, unaligned data

Arguments

out_data: Out data buffer

data_size: Out data size

Safety

The caller is responsible for providing a pointer valid to write out_size bytes

pub unsafe fn read_raw( &mut self, out_data: *mut u8, data_size: usize, ) -> Result<()>

Reads raw (aligned) data

This essentially aligns up the read size to a 4-byte align

Arguments

out_data: Out data buffer

data_size: Out data size

Safety

The caller is responsible for providing a pointer valid to write out_size bytes

pub unsafe fn write_raw_unaligned( &mut self, data: *const u8, data_size: usize, ) -> Result<()>

Writes raw, unaligned data

Arguments

data: In data buffer

data_size: In data size

Safety

The caller is responsible for providing a pointer valid to read out_size bytes

pub fn write_reserve_raw(&mut self, data_size: usize) -> Result<*mut u8>

Reserves a certain (aligned) size at the payload, to be written later (returning the corresponding buffer)

Arguments

data_size: Out data size

pub unsafe fn write_raw( &mut self, data: *const u8, data_size: usize, ) -> Result<()>

Writes raw (aligned) data

This essentially aligns up the write size to a 4-byte alignment in the buffer

Arguments

data: In data buffer

data_size: In data size

Safety

The caller must provide a pointer valid for reading data_size bytes

pub fn write_unaligned<T>(&mut self, t: T) -> Result<()>

Writes an unaligned value

Arguments

t: The value

pub fn write<T: Copy>(&mut self, t: T) -> Result<()>

Writes a value (aligned)

Arguments

t: The value

pub fn read_unaligned<T>(&mut self) -> Result<T>

Reads an unaligned value

pub fn read<T: Copy>(&mut self) -> Result<T>

Reads a value (aligned to a 4-byte word)

pub fn write_str(&mut self, string: &str) -> Result<()>

Writes a string

Note that strings are internally (de)serialized as NUL-terminated UTF-16

Arguments

string: The string to write

pub fn write_interface_token(&mut self, token: &str) -> Result<()>

Writes an interface token

Arguments

token: The interface token name

pub unsafe fn read_sized_raw( &mut self, out_data: *mut u8, out_size: usize, ) -> Result<usize>

Reads raw sized data

For sized data, the data is preceded by its size

Arguments

out_data: Out data buffer

out_size: The maximum size in bytes that can be read into out_data

Safety

The caller is responsible for providing a pointer valid to read out_size bytes

pub fn read_sized<T: Default + Copy>(&mut self) -> Result<T>

Reads a value as sized data

This verifies that the read data is at least big enough to contain the value type, returning ResultReadSizeMismatch otherwise

pub unsafe fn write_sized_raw( &mut self, data: *const u8, data_size: usize, ) -> Result<()>

Writes raw sized data

Arguments

data: In data buffer

data size: In data size

Safety

The caller must provide a pointer valid for reading data_size bytes

pub fn write_sized<T: Copy>(&mut self, t: T) -> Result<()>

Writes a value as sized data

Arguments

t: The value to write

pub fn load_from(&mut self, payload: ParcelPayload)

Loads an external payload in this Parcel

Arguments

payload: The payload

pub fn end_write(&mut self) -> Result<(ParcelPayload, usize)>

Finishes writing and produces the payload

Essentially populates the payload header and returns the current payload, along with its size