nx::hbl
Struct Version

#[repr(C)]
pub struct Version { /* private fields */ }

Represents the hbl-ABI format of the system version

Implementations

impl Version

pub const ATMOSPHERE_OS_IMPL_MAGIC: u64 = 4_707_472_513_604_536_402u64

Represents the [u64] magic present in AbiConfigEntryKey::HosVersion entry values if Atmosphere is currently running

pub const IS_ATMOSPHERE_BIT: u32 = 2_147_483_648u32

Represents the bit set in the Version value if Atmosphere is the current OS implementation

pub const fn empty() -> Self

Creates an empty Version, whose value will be 0.0.0

pub const fn new(value: u32, os_impl_magic: u64) -> Self

Creates a Version from a raw value and the magic representing the current OS implementation

Arguments

value: The raw value

os_impl_magic: The magic value

pub const fn get_major(&self) -> u8

Gets the major component of the Version

pub const fn get_minor(&self) -> u8

Gets the minor component of the Version

pub const fn get_micro(&self) -> u8

Gets the micro component of the Version

pub const fn is_atmosphere(&self) -> bool

Gets whether Atmosphere is the current OS implementation

pub const fn to_version(&self) -> Version

Gets a Version type from this Version