nx::ipc

Module sf

Modules

applet

dispdrv

fatal

fsp

hid

hipc

ldr

lm

lr

mii

ncm

nfp

nv

pm

psc

psm

set

sm

spl

usb

vi

Structs

AppletResourceUserId

AppletResourceUserIds are restricted to the values of zero, or the process’ PID. When they are sent over an IPC interface, they also trigger the sending of a PID descriptor in the HIPC request, so there is an additional field for the PID. This field is filled in by the kernel during a request, and is read out of the headers in the same way as the ProcessIdProcessId above.

Buffer

Handle

ProcessId

Session

Type Aliases

CopyHandle

InAutoSelectBuffer

InFixedPointerBuffer

InMapAliasBuffer

InNonSecureMapAliasBuffer

InPointerBuffer

MoveHandle

OutAutoSelectBuffer

OutFixedPointerBuffer

OutMapAliasBuffer

OutNonSecureMapAliasBuffer

OutPointerBuffer

Unions

EnumAsPrimitiveType

Attribute Macros

ipc_trait

Derive Macros

Request

This creates the required trait implementations for the type to be used as an IPC request parameter. As the type is directly copied into the buffer from an &Self, this will only work on Copy types.

Response

This creates the required trait implementations for the type to be used as an IPC response parameter. As the type is directly copied into the buffer from an &Self, this will only work on Copy types.