nx::ipc::sf
Module applet

Re-exports

pub use super::AppletResourceUserId;

Structs

AllSystemAppletProxiesService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

AppletAttribute

LibraryAppletAccessor

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

LibraryAppletCreator

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

LibraryAppletProxy

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SelfController

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

Storage

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

StorageAccessor

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

WindowController

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

Enums

AppletId

LibraryAppletMode

ScreenShotPermission

Traits

IAllSystemAppletProxiesService

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IAllSystemAppletProxiesServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ILibraryAppletAccessor

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILibraryAppletAccessorServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ILibraryAppletCreator

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILibraryAppletCreatorServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ILibraryAppletProxy

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILibraryAppletProxyServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISelfController

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISelfControllerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IStorage

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IStorageAccessor

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IStorageAccessorServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IStorageServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IWindowController

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IWindowControllerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.