nx::ipc::sf::applet

Trait IWindowControllerServer

pub trait IWindowControllerServer: ISessionObject + Sync {
    // Required methods
    fn get_applet_resource_user_id(&mut self) -> Result<u64>;
    fn acquire_foreground_rights(&mut self) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }
}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn get_applet_resource_user_id(&mut self) -> Result<u64>

fn acquire_foreground_rights(&mut self) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.