nx::ipc::sf::fsp::srv

Trait IFileSystemProxyServer

pub trait IFileSystemProxyServer: ISessionObject + Sync {
    // Required methods
    fn set_current_process(&mut self, process_id: ProcessId) -> Result<()>;
    fn open_sd_card_filesystem(&mut self) -> Result<FileSystem>;
    fn output_access_log_to_sd_card(
        &mut self,
        log_buf: InMapAliasBuffer<u8>,
    ) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }
}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn set_current_process(&mut self, process_id: ProcessId) -> Result<()>

fn open_sd_card_filesystem(&mut self) -> Result<FileSystem>

fn output_access_log_to_sd_card( &mut self, log_buf: InMapAliasBuffer<u8>, ) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.