nx::ipc::sf::fsp

Trait IFileSystemServer

pub trait IFileSystemServer: ISessionObject + Sync {

    // Required methods
    fn create_file(
        &mut self,
        attribute: FileAttribute,
        size: usize,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn delete_file(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn create_directory(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn delete_directory(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn delete_directory_recursively(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn rename_file(
        &mut self,
        old_path_buf: InFixedPointerBuffer<Path>,
        new_path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn rename_directory(
        &mut self,
        old_path_buf: InFixedPointerBuffer<Path>,
        new_path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn get_entry_type(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<DirectoryEntryType>;
    fn open_file(
        &mut self,
        mode: FileOpenMode,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<impl IFileServer + 'static>;
    fn open_directory(
        &mut self,
        mode: DirectoryOpenMode,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<impl IDirectoryServer + 'static>;
    fn commit(&mut self) -> Result<()>;
    fn get_free_space_size(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<usize>;
    fn get_total_space_size(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<usize>;
    fn clean_directory_recursively(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<()>;
    fn get_file_time_stamp_raw(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
    ) -> Result<FileTimeStampRaw>;
    fn query_entry(
        &mut self,
        path_buf: InFixedPointerBuffer<Path>,
        query_id: QueryId,
        in_buf: InNonSecureMapAliasBuffer<u8>,
        out_buf: OutNonSecureMapAliasBuffer<u8>,
    ) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }

}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn create_file( &mut self, attribute: FileAttribute, size: usize, path_buf: InFixedPointerBuffer<Path>, ) -> Result<()>

fn delete_file(&mut self, path_buf: InFixedPointerBuffer<Path>) -> Result<()>

fn create_directory( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<()>

fn delete_directory( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<()>

fn delete_directory_recursively( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<()>

fn rename_file( &mut self, old_path_buf: InFixedPointerBuffer<Path>, new_path_buf: InFixedPointerBuffer<Path>, ) -> Result<()>

fn rename_directory( &mut self, old_path_buf: InFixedPointerBuffer<Path>, new_path_buf: InFixedPointerBuffer<Path>, ) -> Result<()>

fn get_entry_type( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<DirectoryEntryType>

fn open_file( &mut self, mode: FileOpenMode, path_buf: InFixedPointerBuffer<Path>, ) -> Result<impl IFileServer + 'static>

fn open_directory( &mut self, mode: DirectoryOpenMode, path_buf: InFixedPointerBuffer<Path>, ) -> Result<impl IDirectoryServer + 'static>

fn commit(&mut self) -> Result<()>

fn get_free_space_size( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<usize>

fn get_total_space_size( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<usize>

fn clean_directory_recursively( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<()>

fn get_file_time_stamp_raw( &mut self, path_buf: InFixedPointerBuffer<Path>, ) -> Result<FileTimeStampRaw>

fn query_entry( &mut self, path_buf: InFixedPointerBuffer<Path>, query_id: QueryId, in_buf: InNonSecureMapAliasBuffer<u8>, out_buf: OutNonSecureMapAliasBuffer<u8>, ) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.