nx::ipc::sf::hid::shmem

Enum SharedMemoryFormat

pub enum SharedMemoryFormat {
    V1(&'static SharedMemoryFormatV1),
    V2(&'static SharedMemoryFormatV2),
    V3(&'static SharedMemoryFormatV3),
    V4(&'static SharedMemoryFormatV4),
    V5(&'static SharedMemoryFormatV5),
    V6(&'static SharedMemoryFormatV6),
}

Variants

V1(&'static SharedMemoryFormatV1)

V2(&'static SharedMemoryFormatV2)

V3(&'static SharedMemoryFormatV3)

V4(&'static SharedMemoryFormatV4)

V5(&'static SharedMemoryFormatV5)

V6(&'static SharedMemoryFormatV6)

Implementations

impl SharedMemoryFormat

pub unsafe fn from_shmem_ptr(ptr: *const u8) -> Result<Self>

Constructs an enum to get the controller parameters from the shared memory map.

Args

ptr - a const pointer to the controller’s shared memory

Safety

It is the caller’s responsibility to make sure the returned struct does not outlive the shared memory mapping.

pub fn as_ptr(&self) -> *const u8