nx::ipc::sf::hid

Trait IHidSysServer

pub trait IHidSysServer: ISessionObject + Sync {
    // Required methods
    fn send_keyboard_lock_key_event(
        &mut self,
        flags: LockKeyFlags,
    ) -> Result<()>;
    fn acquire_home_button_event_handle(
        &mut self,
        aruid: AppletResourceUserId,
    ) -> Result<CopyHandle>;
    fn activate_home_button(
        &mut self,
        aruid: AppletResourceUserId,
    ) -> Result<()>;
    fn acquire_sleep_button_event_handle(
        &mut self,
        aruid: AppletResourceUserId,
    ) -> Result<CopyHandle>;
    fn activate_sleep_button(
        &mut self,
        aruid: AppletResourceUserId,
    ) -> Result<()>;
    fn acquire_capture_button_event_handle(
        &mut self,
        aruid: AppletResourceUserId,
    ) -> Result<CopyHandle>;
    fn activate_capture_button(
        &mut self,
        aruid: AppletResourceUserId,
    ) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }
}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn send_keyboard_lock_key_event(&mut self, flags: LockKeyFlags) -> Result<()>

fn acquire_home_button_event_handle( &mut self, aruid: AppletResourceUserId, ) -> Result<CopyHandle>

fn activate_home_button(&mut self, aruid: AppletResourceUserId) -> Result<()>

fn acquire_sleep_button_event_handle( &mut self, aruid: AppletResourceUserId, ) -> Result<CopyHandle>

fn activate_sleep_button(&mut self, aruid: AppletResourceUserId) -> Result<()>

fn acquire_capture_button_event_handle( &mut self, aruid: AppletResourceUserId, ) -> Result<CopyHandle>

fn activate_capture_button(&mut self, aruid: AppletResourceUserId) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.