nx::ipc::sf::ldr
Trait IShellInterfaceServer

pub trait IShellInterfaceServer: ISessionObject + Sync {
    // Required methods
    fn set_program_argument_deprecated(
        &mut self,
        program_id: ProgramId,
        args_size: u32,
        args_buf: InPointerBuffer<u8>,
    ) -> Result<()>;
    fn set_program_argument(
        &mut self,
        program_id: ProgramId,
        args_buf: InPointerBuffer<u8>,
    ) -> Result<()>;
    fn flush_arguments(&mut self) -> Result<()>;
    fn atmosphere_register_external_code(
        &mut self,
        program_id: ProgramId,
    ) -> Result<MoveHandle>;
    fn atmosphere_unregister_external_code(
        &mut self,
        program_id: ProgramId,
    ) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }
}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn set_program_argument_deprecated( &mut self, program_id: ProgramId, args_size: u32, args_buf: InPointerBuffer<u8>, ) -> Result<()>

fn set_program_argument( &mut self, program_id: ProgramId, args_buf: InPointerBuffer<u8>, ) -> Result<()>

fn flush_arguments(&mut self) -> Result<()>

fn atmosphere_register_external_code( &mut self, program_id: ProgramId, ) -> Result<MoveHandle>

fn atmosphere_unregister_external_code( &mut self, program_id: ProgramId, ) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.