nx::ipc::sf

Module lm

Structs

LogDestination

Logger

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

Traits

ILoggerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILoggerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ILoggingClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILoggingServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.