nx::ipc::sf

Module mii

Structs

CharInfo

CoreData

CoreDataElementInfo

DatabaseService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SourceFlag

StaticService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

StoreData

Enums

Age

BeardType

CoreDataElement

EyeType

EyebrowType

FaceColor

FacelineColor

FacelineMake

FacelineType

FacelineWrinkle

FontRegion

Gender

GlassType

HairFlip

HairType

MoleType

MouthType

MustacheType

NoseType

SpecialKeyCode

Traits

IDatabaseServiceClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IDatabaseServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IStaticServiceClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IStaticServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Functions

compute_crc16

get_device_id

get_element_info

Type Aliases

CommonColor

CreateId