nx::ipc::sf::ncm
Trait IContentMetaDatabaseServer

pub trait IContentMetaDatabaseServer: ISessionObject + Sync {

    // Required methods
    fn set(
        &mut self,
        meta_key: ContentMetaKey,
        in_rec_buf: InMapAliasBuffer<u8>,
    ) -> Result<()>;
    fn get(
        &mut self,
        meta_key: ContentMetaKey,
        out_rec_buf: OutMapAliasBuffer<u8>,
    ) -> Result<usize>;
    fn remove(&mut self, meta_key: ContentMetaKey) -> Result<()>;
    fn get_content_id_by_type(
        &mut self,
        meta_key: ContentMetaKey,
        cnt_type: ContentType,
    ) -> Result<ContentId>;
    fn list_content_info(
        &mut self,
        out_rec_buf: OutMapAliasBuffer<ContentInfo>,
        meta_key: ContentMetaKey,
        offset: u32,
    ) -> Result<u32>;
    fn list(
        &mut self,
        out_meta_keys: OutMapAliasBuffer<ContentMetaKey>,
        meta_type: ContentMetaType,
        program_id: ProgramId,
        program_id_min: ProgramId,
        program_id_max: ProgramId,
        install_type: ContentInstallType,
    ) -> Result<(u32, u32)>;
    fn get_latest_content_meta_key(
        &mut self,
        program_id: ProgramId,
    ) -> Result<ContentMetaKey>;
    fn list_application(
        &mut self,
        out_app_meta_keys: OutMapAliasBuffer<ApplicationContentMetaKey>,
        meta_type: ContentMetaType,
    ) -> Result<(u32, u32)>;
    fn has(&mut self, meta_key: ContentMetaKey) -> Result<bool>;
    fn has_all(
        &mut self,
        meta_keys_buf: InMapAliasBuffer<ContentMetaKey>,
    ) -> Result<bool>;
    fn get_size(&mut self, meta_key: ContentMetaKey) -> Result<usize>;
    fn get_required_system_version(
        &mut self,
        meta_key: ContentMetaKey,
    ) -> Result<u32>;
    fn get_patch_content_meta_id(
        &mut self,
        meta_key: ContentMetaKey,
    ) -> Result<ProgramId>;
    fn disable_forcibly(&mut self) -> Result<()>;
    fn lookup_orphan_content(
        &mut self,
        content_ids_buf: InMapAliasBuffer<ContentId>,
        out_orphaned_buf: OutMapAliasBuffer<bool>,
    ) -> Result<()>;
    fn commit(&mut self) -> Result<()>;
    fn has_content(
        &mut self,
        meta_key: ContentMetaKey,
        id: ContentId,
    ) -> Result<bool>;
    fn list_content_meta_info(
        &mut self,
        out_meta_infos: OutMapAliasBuffer<ContentMetaInfo>,
        meta_key: ContentMetaKey,
        offset: u32,
    ) -> Result<u32>;
    fn get_attributes(&mut self, meta_key: ContentMetaKey) -> Result<u8>;
    fn get_required_application_version(
        &mut self,
        meta_key: ContentMetaKey,
    ) -> Result<u32>;
    fn get_content_id_by_type_and_offset(
        &mut self,
        meta_key: ContentMetaKey,
        cnt_type: ContentType,
        id_offset: u8,
    ) -> Result<ContentId>;
    fn get_count(&mut self) -> Result<u32>;
    fn get_owner_application_id(
        &mut self,
        meta_key: ContentMetaKey,
    ) -> Result<ApplicationId>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }

}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn set( &mut self, meta_key: ContentMetaKey, in_rec_buf: InMapAliasBuffer<u8>, ) -> Result<()>

fn get( &mut self, meta_key: ContentMetaKey, out_rec_buf: OutMapAliasBuffer<u8>, ) -> Result<usize>

fn remove(&mut self, meta_key: ContentMetaKey) -> Result<()>

fn get_content_id_by_type( &mut self, meta_key: ContentMetaKey, cnt_type: ContentType, ) -> Result<ContentId>

fn list_content_info( &mut self, out_rec_buf: OutMapAliasBuffer<ContentInfo>, meta_key: ContentMetaKey, offset: u32, ) -> Result<u32>

fn list( &mut self, out_meta_keys: OutMapAliasBuffer<ContentMetaKey>, meta_type: ContentMetaType, program_id: ProgramId, program_id_min: ProgramId, program_id_max: ProgramId, install_type: ContentInstallType, ) -> Result<(u32, u32)>

fn get_latest_content_meta_key( &mut self, program_id: ProgramId, ) -> Result<ContentMetaKey>

fn list_application( &mut self, out_app_meta_keys: OutMapAliasBuffer<ApplicationContentMetaKey>, meta_type: ContentMetaType, ) -> Result<(u32, u32)>

fn has(&mut self, meta_key: ContentMetaKey) -> Result<bool>

fn has_all( &mut self, meta_keys_buf: InMapAliasBuffer<ContentMetaKey>, ) -> Result<bool>

fn get_size(&mut self, meta_key: ContentMetaKey) -> Result<usize>

fn get_required_system_version( &mut self, meta_key: ContentMetaKey, ) -> Result<u32>

fn get_patch_content_meta_id( &mut self, meta_key: ContentMetaKey, ) -> Result<ProgramId>

fn disable_forcibly(&mut self) -> Result<()>

fn lookup_orphan_content( &mut self, content_ids_buf: InMapAliasBuffer<ContentId>, out_orphaned_buf: OutMapAliasBuffer<bool>, ) -> Result<()>

fn commit(&mut self) -> Result<()>

fn has_content( &mut self, meta_key: ContentMetaKey, id: ContentId, ) -> Result<bool>

fn list_content_meta_info( &mut self, out_meta_infos: OutMapAliasBuffer<ContentMetaInfo>, meta_key: ContentMetaKey, offset: u32, ) -> Result<u32>

fn get_attributes(&mut self, meta_key: ContentMetaKey) -> Result<u8>

fn get_required_application_version( &mut self, meta_key: ContentMetaKey, ) -> Result<u32>

fn get_content_id_by_type_and_offset( &mut self, meta_key: ContentMetaKey, cnt_type: ContentType, id_offset: u8, ) -> Result<ContentId>

fn get_count(&mut self) -> Result<u32>

fn get_owner_application_id( &mut self, meta_key: ContentMetaKey, ) -> Result<ApplicationId>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.