nx::ipc::sf

Module nfp

Modules

rc

Structs

AdminInfo

AdminInfoFlags

CommonInfo

Date

Debug

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

DebugManager

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

DeviceHandle

McuVersionData

ModelInfo

NfpData

RegisterInfo

RegisterInfoPrivate

System

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SystemManager

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

TagInfo

User

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

UserManager

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

Enums

BreakType

ConsoleFamily

DeviceState

ModelType

MountTarget

State

WriteType

Traits

IDebugClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IDebugManagerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IDebugManagerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IDebugServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISystemClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISystemManagerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISystemManagerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISystemServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IUserClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IUserManagerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IUserManagerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IUserServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Type Aliases

AccessId