nx::ipc::sf::nfp

Trait IUserServer

pub trait IUserServer: ISessionObject + Sync {

    // Required methods
    fn initialize(
        &mut self,
        aruid: AppletResourceUserId,
        mcu_data: InMapAliasBuffer<McuVersionData>,
    ) -> Result<()>;
    fn initialize_2(
        &mut self,
        aruid: AppletResourceUserId,
        mcu_data: InMapAliasBuffer<McuVersionData>,
    ) -> Result<()>;
    fn finalize(&mut self) -> Result<()>;
    fn list_devices(
        &mut self,
        out_devices: OutPointerBuffer<DeviceHandle>,
    ) -> Result<u32>;
    fn start_detection(&mut self, device_handle: DeviceHandle) -> Result<()>;
    fn stop_detection(&mut self, device_handle: DeviceHandle) -> Result<()>;
    fn mount(
        &mut self,
        device_handle: DeviceHandle,
        model_type: ModelType,
        mount_target: MountTarget,
    ) -> Result<()>;
    fn unmount(&mut self, device_handle: DeviceHandle) -> Result<()>;
    fn open_application_area(
        &mut self,
        device_handle: DeviceHandle,
        access_id: AccessId,
    ) -> Result<()>;
    fn get_application_area(
        &mut self,
        device_handle: DeviceHandle,
        out_data: OutMapAliasBuffer<u8>,
    ) -> Result<u32>;
    fn set_application_area(
        &mut self,
        device_handle: DeviceHandle,
        data: InMapAliasBuffer<u8>,
    ) -> Result<()>;
    fn flush(&mut self, device_handle: DeviceHandle) -> Result<()>;
    fn restore(&mut self, device_handle: DeviceHandle) -> Result<()>;
    fn create_application_area(
        &mut self,
        device_handle: DeviceHandle,
        access_id: AccessId,
        data: InMapAliasBuffer<u8>,
    ) -> Result<()>;
    fn get_tag_info(
        &mut self,
        device_handle: DeviceHandle,
        out_tag_info: OutFixedPointerBuffer<TagInfo>,
    ) -> Result<()>;
    fn get_register_info(
        &mut self,
        device_handle: DeviceHandle,
        out_register_info: OutFixedPointerBuffer<RegisterInfo>,
    ) -> Result<()>;
    fn get_common_info(
        &mut self,
        device_handle: DeviceHandle,
        out_common_info: OutFixedPointerBuffer<CommonInfo>,
    ) -> Result<()>;
    fn get_model_info(
        &mut self,
        device_handle: DeviceHandle,
        out_model_info: OutFixedPointerBuffer<ModelInfo>,
    ) -> Result<()>;
    fn attach_activate_event(
        &mut self,
        device_handle: DeviceHandle,
    ) -> Result<CopyHandle>;
    fn attach_deactivate_event(
        &mut self,
        device_handle: DeviceHandle,
    ) -> Result<CopyHandle>;
    fn get_state(&mut self) -> Result<State>;
    fn get_device_state(
        &mut self,
        device_handle: DeviceHandle,
    ) -> Result<DeviceState>;
    fn get_npad_id(&mut self, device_handle: DeviceHandle) -> Result<NpadIdType>;
    fn get_application_area_size(
        &mut self,
        device_handle: DeviceHandle,
    ) -> Result<u32>;
    fn attach_availability_change_event(&mut self) -> Result<CopyHandle>;
    fn recreate_application_area(
        &mut self,
        device_handle: DeviceHandle,
        access_id: AccessId,
        data: InMapAliasBuffer<u8>,
    ) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }

}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn initialize( &mut self, aruid: AppletResourceUserId, mcu_data: InMapAliasBuffer<McuVersionData>, ) -> Result<()>

fn initialize_2( &mut self, aruid: AppletResourceUserId, mcu_data: InMapAliasBuffer<McuVersionData>, ) -> Result<()>

fn finalize(&mut self) -> Result<()>

fn list_devices( &mut self, out_devices: OutPointerBuffer<DeviceHandle>, ) -> Result<u32>

fn start_detection(&mut self, device_handle: DeviceHandle) -> Result<()>

fn stop_detection(&mut self, device_handle: DeviceHandle) -> Result<()>

fn mount( &mut self, device_handle: DeviceHandle, model_type: ModelType, mount_target: MountTarget, ) -> Result<()>

fn unmount(&mut self, device_handle: DeviceHandle) -> Result<()>

fn open_application_area( &mut self, device_handle: DeviceHandle, access_id: AccessId, ) -> Result<()>

fn get_application_area( &mut self, device_handle: DeviceHandle, out_data: OutMapAliasBuffer<u8>, ) -> Result<u32>

fn set_application_area( &mut self, device_handle: DeviceHandle, data: InMapAliasBuffer<u8>, ) -> Result<()>

fn flush(&mut self, device_handle: DeviceHandle) -> Result<()>

fn restore(&mut self, device_handle: DeviceHandle) -> Result<()>

fn create_application_area( &mut self, device_handle: DeviceHandle, access_id: AccessId, data: InMapAliasBuffer<u8>, ) -> Result<()>

fn get_tag_info( &mut self, device_handle: DeviceHandle, out_tag_info: OutFixedPointerBuffer<TagInfo>, ) -> Result<()>

fn get_register_info( &mut self, device_handle: DeviceHandle, out_register_info: OutFixedPointerBuffer<RegisterInfo>, ) -> Result<()>

fn get_common_info( &mut self, device_handle: DeviceHandle, out_common_info: OutFixedPointerBuffer<CommonInfo>, ) -> Result<()>

fn get_model_info( &mut self, device_handle: DeviceHandle, out_model_info: OutFixedPointerBuffer<ModelInfo>, ) -> Result<()>

fn attach_activate_event( &mut self, device_handle: DeviceHandle, ) -> Result<CopyHandle>

fn attach_deactivate_event( &mut self, device_handle: DeviceHandle, ) -> Result<CopyHandle>

fn get_state(&mut self) -> Result<State>

fn get_device_state( &mut self, device_handle: DeviceHandle, ) -> Result<DeviceState>

fn get_npad_id(&mut self, device_handle: DeviceHandle) -> Result<NpadIdType>

fn get_application_area_size( &mut self, device_handle: DeviceHandle, ) -> Result<u32>

fn attach_availability_change_event(&mut self) -> Result<CopyHandle>

fn recreate_application_area( &mut self, device_handle: DeviceHandle, access_id: AccessId, data: InMapAliasBuffer<u8>, ) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.