nx::ipc::sf::usb

Module hs

Structs

ClientEpSession

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

ClientIfSession

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

ClientRootSession

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

DeviceFilter

DeviceFilterFlags

InterfaceInfo

InterfaceProfile

InterfaceQueryOutput

XferReport

Enums

InterfaceAvailableEventId

Traits

IClientEpSession

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IClientEpSessionServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IClientIfSession

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IClientIfSessionServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IClientRootSession

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IClientRootSessionServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.