nx::ipc::sf::usb::hs

Trait IClientEpSessionServer

pub trait IClientEpSessionServer: ISessionObject + Sync {

    // Required methods
    fn submit_out_request(
        &mut self,
        size: u32,
        unk: u32,
        buf: InMapAliasBuffer<u8>,
    ) -> Result<u32>;
    fn re_open(&mut self) -> Result<()>;
    fn submit_in_request(
        &mut self,
        size: u32,
        unk: u32,
        out_buf: OutMapAliasBuffer<u8>,
    ) -> Result<u32>;
    fn close(&mut self) -> Result<()>;
    fn reset(&mut self) -> Result<()>;
    fn get_completion_event(&mut self) -> Result<CopyHandle>;
    fn close_deprecated(&mut self) -> Result<()>;
    fn populate_ring(&mut self) -> Result<()>;
    fn post_buffer_async(
        &mut self,
        size: u32,
        buf_addr: u64,
        unk: u64,
    ) -> Result<u32>;
    fn get_xfer_report_deprecated(
        &mut self,
        count: u32,
        out_reports_buf: OutMapAliasBuffer<XferReport>,
    ) -> Result<u32>;
    fn get_xfer_report(
        &mut self,
        count: u32,
        out_reports_buf: OutAutoSelectBuffer<XferReport>,
    ) -> Result<u32>;
    fn batch_buffer_async_deprecated(
        &mut self,
        urb_count: u32,
        unk_1: u32,
        unk_2: u32,
        buf_addr: u64,
        unk_3: u64,
        urb_sizes_buf: InMapAliasBuffer<u32>,
    ) -> Result<u32>;
    fn batch_buffer_async(
        &mut self,
        urb_count: u32,
        unk_1: u32,
        unk_2: u32,
        buf_addr: u64,
        unk_3: u64,
        urb_sizes_buf: InAutoSelectBuffer<u32>,
    ) -> Result<u32>;
    fn create_smmu_space(&mut self, unk: [u8; 16]) -> Result<()>;
    fn share_report_ring(
        &mut self,
        unk: [u8; 4],
        unk_handle: CopyHandle,
    ) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }

}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn submit_out_request( &mut self, size: u32, unk: u32, buf: InMapAliasBuffer<u8>, ) -> Result<u32>

fn re_open(&mut self) -> Result<()>

fn submit_in_request( &mut self, size: u32, unk: u32, out_buf: OutMapAliasBuffer<u8>, ) -> Result<u32>

fn close(&mut self) -> Result<()>

fn reset(&mut self) -> Result<()>

fn get_completion_event(&mut self) -> Result<CopyHandle>

fn close_deprecated(&mut self) -> Result<()>

fn populate_ring(&mut self) -> Result<()>

fn post_buffer_async( &mut self, size: u32, buf_addr: u64, unk: u64, ) -> Result<u32>

fn get_xfer_report_deprecated( &mut self, count: u32, out_reports_buf: OutMapAliasBuffer<XferReport>, ) -> Result<u32>

fn get_xfer_report( &mut self, count: u32, out_reports_buf: OutAutoSelectBuffer<XferReport>, ) -> Result<u32>

fn batch_buffer_async_deprecated( &mut self, urb_count: u32, unk_1: u32, unk_2: u32, buf_addr: u64, unk_3: u64, urb_sizes_buf: InMapAliasBuffer<u32>, ) -> Result<u32>

fn batch_buffer_async( &mut self, urb_count: u32, unk_1: u32, unk_2: u32, buf_addr: u64, unk_3: u64, urb_sizes_buf: InAutoSelectBuffer<u32>, ) -> Result<u32>

fn create_smmu_space(&mut self, unk: [u8; 16]) -> Result<()>

fn share_report_ring( &mut self, unk: [u8; 4], unk_handle: CopyHandle, ) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.