nx::ipc::sf::usb::hs
Trait IClientRootSessionServer

pub trait IClientRootSessionServer: ISessionObject + Sync {

    // Required methods
    fn bind_client_process(
        &mut self,
        self_process_handle: CopyHandle,
    ) -> Result<()>;
    fn query_all_interfaces_deprecated(
        &mut self,
        filter: DeviceFilter,
        out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>,
    ) -> Result<u32>;
    fn query_all_interfaces(
        &mut self,
        filter: DeviceFilter,
        out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>,
    ) -> Result<u32>;
    fn query_available_interfaces_deprecated(
        &mut self,
        filter: DeviceFilter,
        out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>,
    ) -> Result<u32>;
    fn query_available_interfaces(
        &mut self,
        filter: DeviceFilter,
        out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>,
    ) -> Result<u32>;
    fn query_acquired_interfaces_deprecated(
        &mut self,
        out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>,
    ) -> Result<u32>;
    fn query_acquired_interfaces(
        &mut self,
        out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>,
    ) -> Result<u32>;
    fn create_interface_available_event_deprecated(
        &mut self,
        event_id: InterfaceAvailableEventId,
        filter: DeviceFilter,
    ) -> Result<CopyHandle>;
    fn create_interface_available_event(
        &mut self,
        event_id: InterfaceAvailableEventId,
        filter: DeviceFilter,
    ) -> Result<CopyHandle>;
    fn destroy_interface_available_event_deprecated(
        &mut self,
        event_id: InterfaceAvailableEventId,
    ) -> Result<()>;
    fn destroy_interface_available_event(
        &mut self,
        event_id: InterfaceAvailableEventId,
    ) -> Result<()>;
    fn get_interface_state_change_event_deprecated(
        &mut self,
    ) -> Result<CopyHandle>;
    fn get_interface_state_change_event(&mut self) -> Result<CopyHandle>;
    fn acquire_usb_if_deprecated(
        &mut self,
        id: u32,
        out_profile_buf: OutMapAliasBuffer<InterfaceProfile>,
    ) -> Result<ClientIfSession>;
    fn acquire_usb_if(
        &mut self,
        id: u32,
        out_info_buf: OutMapAliasBuffer<InterfaceInfo>,
        out_profile_buf: OutMapAliasBuffer<InterfaceProfile>,
    ) -> Result<ClientIfSession>;
    fn get_descriptor_string(
        &mut self,
        unk_1: u8,
        unk_2: bool,
        unk_maybe_id: u32,
        out_desc_buf: OutMapAliasBuffer<u8>,
    ) -> Result<u32>;
    fn reset_device(&mut self, unk: u32) -> Result<()>;

    // Provided method
    fn try_handle_request_by_id(
        &mut self,
        req_id: u32,
        protocol: CommandProtocol,
        ctx: &mut ServerContext<'_>,
    ) -> Option<Result<()>> { ... }

}

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Required Methods

fn bind_client_process(&mut self, self_process_handle: CopyHandle) -> Result<()>

fn query_all_interfaces_deprecated( &mut self, filter: DeviceFilter, out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>, ) -> Result<u32>

fn query_all_interfaces( &mut self, filter: DeviceFilter, out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>, ) -> Result<u32>

fn query_available_interfaces_deprecated( &mut self, filter: DeviceFilter, out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>, ) -> Result<u32>

fn query_available_interfaces( &mut self, filter: DeviceFilter, out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>, ) -> Result<u32>

fn query_acquired_interfaces_deprecated( &mut self, out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>, ) -> Result<u32>

fn query_acquired_interfaces( &mut self, out_intfs: OutMapAliasBuffer<InterfaceQueryOutput>, ) -> Result<u32>

fn create_interface_available_event_deprecated( &mut self, event_id: InterfaceAvailableEventId, filter: DeviceFilter, ) -> Result<CopyHandle>

fn create_interface_available_event( &mut self, event_id: InterfaceAvailableEventId, filter: DeviceFilter, ) -> Result<CopyHandle>

fn destroy_interface_available_event_deprecated( &mut self, event_id: InterfaceAvailableEventId, ) -> Result<()>

fn destroy_interface_available_event( &mut self, event_id: InterfaceAvailableEventId, ) -> Result<()>

fn get_interface_state_change_event_deprecated(&mut self) -> Result<CopyHandle>

fn get_interface_state_change_event(&mut self) -> Result<CopyHandle>

fn acquire_usb_if_deprecated( &mut self, id: u32, out_profile_buf: OutMapAliasBuffer<InterfaceProfile>, ) -> Result<ClientIfSession>

fn acquire_usb_if( &mut self, id: u32, out_info_buf: OutMapAliasBuffer<InterfaceInfo>, out_profile_buf: OutMapAliasBuffer<InterfaceProfile>, ) -> Result<ClientIfSession>

fn get_descriptor_string( &mut self, unk_1: u8, unk_2: bool, unk_maybe_id: u32, out_desc_buf: OutMapAliasBuffer<u8>, ) -> Result<u32>

fn reset_device(&mut self, unk: u32) -> Result<()>

Provided Methods

fn try_handle_request_by_id( &mut self, req_id: u32, protocol: CommandProtocol, ctx: &mut ServerContext<'_>, ) -> Option<Result<()>>

The dynamic dispatch function that calls into the IPC server functions. This should only be called from the [$crate::ipc::server::ServerManager] and not from client code. Examples for implementing [ISessionObject][$crate::ipc::server::ISessionObject] or [IMitmServerOject][$crate::ipc::server::IMitmServerObject] can be found in the nx crate.