Function nx::la::launch_wait_library_applet

pub fn launch_wait_library_applet<I: Copy, O: Copy>(
    id: AppletId,
    common_args: CommonArguments,
    input: I
) -> Result<O>

Wrapper to create, launch and wait for a library applet, expecting simple input and output data

The mode used (since all simple library applets expect it) is LibraryAppletMode::AllForeground

Note that this won’t be useful, for instance, with library applets taking interactive in/out data, like AppletId::LibraryAppletSwkbd

Arguments

id: The AppletId of the library applet to create

common_args: The library applet-specific CommonArguments to send as input

input: The only input data to send after the CommonArguments