Struct nx::la::LibraryAppletHolder

pub struct LibraryAppletHolder { /* private fields */ }

Represents a wrapper type for using library applets

Implementations

impl LibraryAppletHolder

pub fn new(accessor: Shared<dyn ILibraryAppletAccessor>) -> Result<Self>

Creates a LibraryAppletHolder from an existing ILibraryAppletAccessor shared object

This shouldn’t be manually created unless the accessor object was obtained manually (see create_library_applet)

pub fn get_accessor(&self) -> Shared<dyn ILibraryAppletAccessor>

Gets the underlying ILibraryAppletAccessor shared object

pub fn push_in_data_storage(
&mut self,
storage: Shared<dyn IStorage>
) -> Result<()>

Pushes an input IStorage shared object to the library applet

pub fn push_in_data<T: Copy>(&mut self, t: T) -> Result<()>

Pushes input data to the library applet

This is a wrapper which creates an IStorage object with the given value and pushes it

pub fn start(&mut self) -> Result<()>

Starts the library applet

pub fn join(&mut self) -> Result<()>

Waits until the library applet’s state-changed event signals

This effectively waits until the library applet exits

pub fn pop_out_data_storage(&mut self) -> Result<Shared<dyn IStorage>>

Pops an output IStorage shared object from the library applet

pub fn pop_out_data<O: Copy>(&mut self) -> Result<O>

Pops output data from the library applet

This is a wrapper which pops an IStorage object and reads its data (reads [size_of][cmem::size_of] O bytes and returns that data)