Module nx::mem::alloc

Allocator implementation and definitions

Modules

rc

Allocation-specific result definitions

Structs

Buffer

Represents a wrapped and manually managed heap value

Layout

Layout of a block of memory.

Constants

PAGE_ALIGNMENT

Traits

Allocator

Represents a heap allocator for this library

Functions

allocate

Allocates heap memory using the global allocator

delete

Deletes a heap value

initialize

Initializes the global allocator with the given address and size

is_enabled

Gets whether heap allocations are enabled

new

Creates a new heap value using the global allocator

release

Releases allocated memory