nx::mem::alloc

Function initialize

pub fn initialize(heap: PointerAndSize) -> bool

Initializes the global allocator with the given address and size. Returns a bool to indicate if the memory was consumed by the allocator

Arguments

heap: The heap address and size