nx::mem::alloc

Struct Buffer

pub struct Buffer<T, A: Allocator = Global> {
    pub ptr: *mut T,
    pub layout: Layout,
    /* private fields */
}

Represents a wrapped and manually managed heap value

Note that a Buffer is able to hold both a single value or an array of values of the provided type

Fields
ptr: *mut T

The actual heap value

layout: Layout

The memory’s layout

Implementations

impl<T> Buffer<T>

pub fn new(align: usize, count: usize) -> Result<Self>

Creates a new Buffer using the global allocator

Arguments

align: The align to use

count: The count of values to allocate

pub const fn empty() -> Self

Creates a new, invalid Buffer

impl<T, A: Allocator> Buffer<T, A>

pub fn new_in(align: usize, count: usize, allocator: A) -> Result<Self>

Creates a new Buffer using a given allocator

Arguments

align: The align to use

count: The count of values to allocate

allocator: The allocator to use

pub fn is_valid(&self) -> bool

Gets whether this Buffer is valid

pub fn into_raw(value: Self) -> *mut [T]

pub unsafe fn release(&mut self)

Releases the Buffer

The Buffer becomes invalid after this

Safety

The buffer must never be read after this, as the internal buffer pointer is wiped. The buffer must also be for