nx::mem::alloc

Trait AllocatorEx

pub unsafe trait AllocatorEx: Allocator {
    // Provided methods
    fn new<T>(&self) -> Result<NonNull<T>> { ... }
    fn delete<T>(&self, t: *mut T) { ... }
}

Represents a heap allocator for this library

Safety

As with the regular Allocator trait, the delete function can only be called on pointers produced by the same implementation’s new

Provided Methods

fn new<T>(&self) -> Result<NonNull<T>>

Allocates a new heap value

fn delete<T>(&self, t: *mut T)

Releases a heap value

The value must have been created using AllocatorEx::new

Arguments

t: Heap value address

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.