nx::rand
Trait RandomGenerator

pub trait RandomGenerator {
    // Required method
    fn random_bytes(&mut self, buf: *mut u8, size: usize) -> Result<()>;

    // Provided method
    fn random<T: Copy + Default>(&mut self) -> Result<T> { ... }
}

Represents a psudo-RNG

Required Methods

fn random_bytes(&mut self, buf: *mut u8, size: usize) -> Result<()>

Fills the given memory region with random bytes

Arguments

buf: Memory region address

size: Memory region size

Provided Methods

fn random<T: Copy + Default>(&mut self) -> Result<T>

Generates a value filled with random contents

This is, of course, meant to be used with types where filling them with random data will be a valid value

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.