nx::result

Struct ResultCode

#[repr(C)]
pub struct ResultCode { /* private fields */ }

Represents a (raw) result value used all over the OS

These are referred as Result on docs/official code, but we intentionally name it as ResultCode to distinguish it from the Result enum type

Results are often displayed/shown, for example, like 2168-0002, which corresponds to <2000 + module>-<description>

[Debug][fmt::Debug] formatting formats the results as a hex-value (0x4A8), while [Display][fmt::Display] formatting formats the result in the format described above (2168-0002)

Implementations

impl ResultCode

pub const fn new(value: u32) -> Self

Creates a ResultCode from a raw value

Arguments

value: The raw value

pub const fn new_err<T>(value: u32) -> Result<T>

Wrapper for creating a new [Result::Err] value with the following raw result

Arguments

value: The raw value, note that it mustnโt be 0/success (that would be undefined behavior)

pub const fn is_success(&self) -> bool

Returns whether the ResultCode is successful

A result value of 0 is a successful value, this essentially checks that

pub const fn is_failure(&self) -> bool

Returns whether the ResultCode is not successful

This is the exact opposite of is_success

pub const fn get_value(&self) -> u32

Gets the raw value of the ResultCode

pub const fn get_module(&self) -> u32

Gets the module of the ResultCode

pub const fn get_description(&self) -> u32

Gets the description of the ResultCode