Module nx::rrt0

Initial code/entrypoint support and utils

Custom entrypoint

If you wish to define your custom entrypoint, you can do so by redefining the __nx_rrt0_entry weak fn.

Example (check here for more entrypoint details):

#[no_mangle]
unsafe extern "C" fn __nx_rrt0_entry(arg0: usize, arg1: usize) {
    // ...
}

Custom version setup

On the default entrypoint routine, the internal system version (see get_version and set_version) gets set the following way:

If the process was launched through HBL, use the “HOS version” value we got from it

Otherwise (and if using the services feature), use settings services (SystemSettingsServer) to get it

If you wish to define your custom version setup (for instance, in contexts in which you wish to avoid executing the aforementioned setup), you can do so by redefining the initialize_version weak fn.

Example:

#[no_mangle]
fn initialize_version(hbl_hos_version_opt: Option<hbl::Version>) {
    // ...
}

Structs

ModulePath

Represents the process module format used by processes

Enums

ExecutableType

Represents the executable type of the current process

Functions

exit

Exits the current process

get_executable_type

Gets the current process’s executable type

get_module_name

Gets this process’s module name

Type Definitions

ExitFn

Represents the fn pointer used for exiting