nx
Module service

Base service/named port support and wrappers

Modules

applet

dispdrv

fatal

fsp

hid

ldr

lm

lr

mii

ncm

nfp

nv

pm

psc

psm

set

sm

spl

usb

vi

Traits

INamedPort

Represents a named port interface

IService

Represents a service interface

Functions

new_named_port_object

Wrapper for connecting to a named port and instantiating the wrapper interface over the specified named port

new_service_object

Wrapper for accessing a service and instantiating the wrapper interface over the specified service