nx::service

Trait INamedPort

pub trait INamedPort: IClientObject {
    // Required methods
    fn get_name() -> &'static str;
    fn post_initialize(&mut self) -> Result<()>;
}

Represents a named port interface

Interfaces which wrap named ports (see manage_named_port or connect_to_named_port) must implement this trait

Required Methods

fn get_name() -> &'static str

Gets the name to be used to connect to the named port (via connect_to_named_port)

fn post_initialize(&mut self) -> Result<()>

This will get executed after connecting to the named port in new_named_port_object, allowing for extra initialization

Some interfaces may have initialization commands (check SM’s case) which can be automatically called this way

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.