Module nx::smc

Secure monitor support and wrappers

Structs

Input

Represents the secure monitor call input layout (special case of Arguments for input)

Output

Represents the secure monitor call output layout (special case of Arguments for output)

Enums

FunctionId

Represents the secure monitor function IDs

Result

Represents the result values returned on secure monitor call responses

Constants

GENERATE_RANDOM_BYTES_MAX_SIZE

Represents the maximum size of the random bytes one can get in the generate_random_bytes SMC

Functions

generate_random_bytes

Secure monitor call which generates random bytes

Type Definitions

Arguments

Represents the raw argument layout used in secure monitor calls