Struct nx::Node

pub struct Node<'a> {
    // some fields omitted
}

A node in an NX file.

Methods

impl<'a> Node<'a>

unsafe fn construct(data: &'a Data, file: &'a File) -> Node<'a>

Creates a Node from the data representing it and the file the data is from.

fn is_empty(&self) -> bool

Gets whether or not the node is empty.

fn name(&self) -> &'a str

Gets the name of this node from the string table.

fn iter(&self) -> Nodes<'a>

Gets an iterator over this node's children.