nx

Module sync

Synchronization support and utils

Modules

sys

Structs

Mutex

Represents a value whose access is controlled by an inner Mutex

MutexGuard

ReadGuard

RwLock

ScopedLock

Represents a type which will lock a given Mutex on creation and unlock it on destruction, effectively guarding it

WriteGuard