Struct nx::sync::Locked

pub struct Locked<T> { /* private fields */ }

Represents a value whose access is controlled by an inner Mutex

Implementations

impl<T> Locked<T>

pub const fn new(is_recursive: bool, t: T) -> Self

Creates a new Locked with a value

Arguments

is_recursive: Whether the inner Mutex is recursive

t: The value to store

pub const fn get_lock(&self) -> &mut Mutex

Gets a reference to the inner Mutex

pub fn get(&self) -> &mut T

Gets a reference of the value, doing a lock-unlock operation in the process

pub fn set(&mut self, t: T)

Sets a value, doing a lock-unlock operation in the process

impl<T: Copy> Locked<T>

pub fn get_val(&self) -> T

Gets a copy of the value, doing a lock-unlock operation in the process