nx
Module thread

Threading support and wrappers

Modules

imp

rc

Thread-specific result definitions

scoped

Structs

Builder

Thread factory, which can be used in order to configure the properties of a new thread.

JoinHandle

An owned permission to join on a thread (block on its termination).

Thread

A handle to a thread.

ThreadLocalRegion

Represents the console’s Thread Local Region layout

Enums

ThreadPriority

Represents the priority of the (thread)Thread that we are spawning

ThreadStartCore

Represents what core the (thread)Thread should start on

ThreadState

Represents the state of a Thread

YieldType

Represents the thread yielding types for cooperative multitasking

Functions

available_parallelism

exit

Exits the current thread

get_thread_local_region⚠

Gets the current thread’s ThreadLocalRegion address

set_current_thread⚠

Sets the current Thread reference on the current ThreadLocalRegion

sleep

Sleeps for the given timeout

spawn

Spawns a new thread, returning a JoinHandle for it.

yield

Type Aliases

Result

A specialized Result type for threads.

ThreadName

Represents a Thread’s name, a 32-byte ArrayString