nx::thread
Function set_current_thread

pub unsafe fn set_current_thread(thread_ref: *mut Thread)

Sets the current Thread reference on the current ThreadLocalRegion

This is internally used when launching a Thread, and probably shouldn’t be used manually

Arguments

thread_ref: The Thread address to set

SAFETY: thread_ref must be a valid Thread pointer that will not move until the thread is finished running.