nx::util

Struct ArrayString

#[repr(C)]
pub struct ArrayString<const S: usize> { /* private fields */ }

Represents a C-like string of a given size (mostly like a C char[S] array)

Note that char is 4-bytes in Rust for encoding reasons, thus we must stick to u8 arrays

Implementations

impl<const S: usize> ArrayString<S>

pub const fn new() -> Self

Creates an empty ArrayString

pub const fn from_raw(raw_bytes: [u8; S]) -> Self

Creates a ArrayString from a given byte array

Arguments

raw_bytes: Byte array to use

pub const fn from_str_truncate_null(string: &str) -> Self

Creates a ArrayString from a given &str

This creates an empty ArrayString and initializes it with the provided string. This will copy at max S - 1 bytes/chars in order to ensure that the string is NUL-terminated. This will truncate the string at the first null, so we can unconditionally return and keep it const.

Arguments

string: The &str to use

pub fn from_str(string: &str) -> Self

Creates a ArrayString from a given &str

This creates an empty ArrayString and calls ArrayString::set_str on it

Arguments

string: The &str to use

pub fn from_string(string: &String) -> Self

Creates a ArrayString from a given String

This creates an empty ArrayString and calls ArrayString::set_string on it

Arguments

string: The String to use

pub fn len(&self) -> usize

Returns the length of the ArrayString

This is similar to Cโs strlen() function, thus taking into account the stringโs NUL-termination

pub fn is_empty(&self) -> bool

Returns whether this ArrayString is empty

pub fn set_str(&mut self, string: &str) -> Result<()>

Sets a &str as the contents of this ArrayString

This will copy at max S - 1 bytes/chars in order to ensure that the string is NUL-terminated Returns and error when the string has internal nulls. Truncates the written strings over S-1 bytes in length.

Arguments

string: The &str to set

pub fn set_string(&mut self, string: &String) -> Result<()>

Sets a string as the contents of this ArrayString

This will copy at max S - 1 bytes/chars in order to ensure that the string is NUL-terminated

Arguments

string: The content to set

pub fn get_str(&self) -> Result<&str>

Gets a &str corresponding to this ArrayString

pub fn get_string(&self) -> Result<String>

Gets a String corresponding to this ArrayString

pub fn as_buffer(&self) -> &[u8; S]

Borrows a view into the whole array

pub fn as_bytes(&self) -> &[u8]

Borrows only the initialized bytes (including the null terminator)