nx::util

Struct ArrayWideString

#[repr(C)]
pub struct ArrayWideString<const S: usize> { /* private fields */ }

Represents a C-like 16-bit string of a given size (mostly like a C char16_t[S] array)

Note that char is 4-bytes in Rust for encoding reasons, thus we must stick to u16 arrays

Implementations

impl<const S: usize> ArrayWideString<S>

pub const fn new() -> Self

Creates an empty ArrayWideString

pub const fn from_raw(raw_bytes: [u16; S]) -> Self

Creates a ArrayWideString from a given byte array

Arguments

raw_bytes: Byte array to use

pub fn from_string(string: String) -> Self

Creates a ArrayWideString from a given String

This creates an empty ArrayWideString and calls ArrayWideString::set_string on it

Arguments

string: The String to use

pub fn len(&self) -> usize

Returns the length of the ArrayWideString

This is similar to Cโs strlen() function, thus taking into account the stringโs NUL-termination

pub fn is_empty(&self) -> bool

Returns if this ArrayWideString is empty

pub fn set_str(&mut self, string: &str)

Sets a &str as the contents of this ArrayWideString

This will copy at max S - 1 bytes/chars in order to ensure that the string is NUL-terminated

Arguments

string: The &str to set

pub fn set_string(&mut self, string: impl AsRef<str>)

Sets a String as the contents of this ArrayWideString

This will copy at max S - 1 bytes/chars in order to ensure that the string is NUL-terminated

Arguments

string: The String to set

pub fn get_string(&self) -> Result<String>

Gets a String corresponding to this ArrayWideString

pub fn as_buffer(&self) -> &[u16; S]

Borrows a view into the whole array

pub fn as_u16s(&self) -> &[u16]

Borrows only the initialized bytes (including the null terminator)