Function nx::version::set_version

pub fn set_version(ver: Version)

Sets the global Version, used in the library as the system Version

This is used on rrt0 to set the actual system version, and shouldn’t be used for other purposes unless you really know what you’re doing

Arguments

ver: The system Version to set globally for the library