nx::version

Struct VersionInterval

pub struct VersionInterval { /* private fields */ }

Represents an interval between versions, being optionally limited on both sides

An interval limited on both sides is, for example, 1.0.0-5.1.0 (inclusive)

An interval limited on one side is, for example, *-3.0.0 (any version lower or equal to 3.0.0) or 2.3.0-* (any version higher or equal to 2.3.0)

Implementations

impl VersionInterval

pub const fn all() -> Self

Creates a non-limited VersionInterval, essentially an interval allowing any versions

pub const fn from(min: Version) -> Self

Creates a left-limited VersionInterval, including any version higher or equal to min

Arguments

min: The minimum Version limiting the interval

pub const fn to(max: Version) -> Self

Creates a right-limited VersionInterval, including any version lower or equal to max

Arguments

max: The maximum Version limiting the interval

pub const fn from_to(min: Version, max: Version) -> Self

Creates a limited VersionInterval, including any version between min and max (inclusive)

Arguments

min: The minimum Version limiting the interval

max: The maximum Version limiting the interval

pub fn contains(&self, ver: Version) -> bool

Returns whether ver is contained in the interval

Arguments

ver: The Version to check