nx

Module vmem

Virtual memory support

Structs

StandardRegions

The standard memory regions for NX processes

VirtualRegion

Represents a virtual region of memory, represented as pointer-sized uints. i.e. [start, end)

Functions

allocate

Finds available virtual memory for the specified size, returning it’s address

get_address_space

Gets the current process’s address space VirtualRegion

get_heap_region

Gets the current process’s heap VirtualRegion

get_legacy_alias_region

Gets the current process’s legacy alias VirtualRegion

get_stack_region

Gets the current process’s stack VirtualRegion

initialize

Initializes virtual memory support