Module nx::wait

Sync/waiting utilities and wrappers

Structs

RemoteEvent

Represents an event via a remote handle

SystemEvent

Represents a system event with server and client handles

Waiter

Represents a waiting object for a handle

Enums

WaiterType

Represents how a waiter operates (essentially, whether it gets automatically cleared after being signaled)

Constants

MAX_OBJECT_COUNT

Represents the max amount of objects the Nintendo Switch kernel can wait-sync on at the same time (like Windows)

Functions

wait

Waits for several Waiters for a specified timeout, returning the index of the waiter which signals first

wait_handles

Waits for several handles for a specified timeout, returning the index of the handle which signals first