nx::wait

Enum WaiterType

pub enum WaiterType {
    Handle,
    HandleWithClear,
}

Represents how a waiter operates (essentially, whether it gets automatically cleared after being signaled)

Variants

Handle

A simple handle, that doesn’t get cleared when the waiter wakes

HandleWithClear

A wait handle that has the signal automatically cleared