nx::wait

Struct Waiter

pub struct Waiter { /* private fields */ }

Represents a waiting object for a handle

Implementations

impl Waiter

pub const fn from(handle: Handle, wait_type: WaiterType) -> Self

Creates a new Waiter from a handle and a type

Arguments

handle - The waiter handle

wait_type - Thr waiter type

pub const fn from_handle(handle: Handle) -> Self

Creates a new Waiter from a handle and WaiterType::Handle type

Arguments

handle - The waiter handle

pub const fn from_handle_with_clear(handle: Handle) -> Self

Creates a new Waiter from a handle and WaiterType::HandleWithClear type

Arguments

handle - The waiter handle