nx
Module elf

ELF (aarch64) support and utils

Modules

mod0

MOD0 format utils

rc

ELF-related result definitions

Structs

Dyn

Represents an ELF dynamic entry.

EhFrameHdrPtr

A stuct containing a pointer sized int, representing a pointer to the start of the eh_frame_hdr elf section. This is obviously not a great option to use with Rust’s upcoming strict/exposed providence APIs, but works fine here as the Switch has a single address space and the memory will have a static lifetime that is longer than the currently running code.

InfoSymbol

Represents an ELF info symbol.

Rel

Represents an ELF Rel type.

Rela

Represents an ELF Rela type

Enums

RelocationType

Represents ELF relocation types.

Tag

Represents ELF tags. Cherry picked from valid relocation types.

Functions

relocate_with_dyn⚠

Relocates a base address with its corresponding Dyn reference

Unions

Info

Represents an ELF info value.