Module nx::fs

FileSystem support

Modules

nca

pfs0

rc

FS-related result definitions

sf

Helper object types to translate from our FS objects to IPC FS interfaces

subdir

Helper object types for treating FS subdirectories as IPC filesystems

Structs

DirectoryAccessor

Represents a wrapper type to simplify directory access

FileAccessor

Represents a wrapper type to simplify file access

FileOpenOption

Represents options for opening files

ProxyDirectory

Represents a wrapper Directory implementation to translate IPC IDirectory objects to Directory objects

ProxyFile

Represents a wrapper File implementation to translate IPC IFile objects to File objects

ProxyFileSystem

Represents a wrapper FileSystem implementation to translate IPC IFileSystem objects to FileSystem objects

Enums

Whence

Represents an offset kind/relativeness

Traits

Directory

Represents a directory

File

Represents a file

FileSystem

Represents a filesystem

Functions

clean_directory_recursively

Deletes all the children files/directories inside a directory

commit

Commits on a filesystem

convert_file_open_mode_to_option

Converts a FileOpenMode to a FileOpenOption

convert_file_open_option_to_mode

Converts a FileOpenOption to a FileOpenMode

create_directory

Creates a directory

create_file

Creates a file

delete_directory

Deletes a directory

delete_directory_recursively

Deletes a directory and all its children files/directories

delete_file

Deletes a file

finalize_fspsrv_session

Finalizes fsp-srv support

format_path

Returns the FileSystem and the processed path corresponding to a given path

get_entry_type

Gets a path’s DirectoryEntryType

get_file_time_stamp_raw

Gets the FileTimeStampRaw of a file

get_free_space_size

Gets the free space size at a given path

get_fspsrv_session

Gets the global IFileSystemProxy shared object used for fsp-srv support

get_path_filesystem

Returns the FileSystem corresponding to a given path

get_total_space_size

Gets the total space size at a given path

initialize_fspsrv_session

Initializes fsp-srv support instantiating a FileSystemProxy shared object

initialize_fspsrv_session_with

Initializes fsp-srv support with a given IFileSystemProxy shared object

is_fspsrv_session_initialized

Gets whether fsp-srv support was initialized

mount

Mounts a FileSystem

mount_fsp_filesystem

Essentially creates a ProxyFileSystem and mounts it

mount_sd_card

Mounts the system’s SD card using fsp-srv support

open_directory

Opens a directory as a DirectoryAccessor

open_file

Opens a file as a FileAccessor

query_entry

Queries on a path

rename

Renames a file/directory

rename_directory

Renames a directory

rename_file

Renames a file

set_concatenation_file_attribute

Sets the “concatenation file” attribute on a file

unmount

Unmounts a mounted filesystem

unmount_all

Unmounts all filesystems

Type Definitions

DirectoryEntry

DirectoryEntryType

DirectoryOpenMode

FileAttribute

FileOpenMode

FileQueryRangeInfo

FileReadOption

FileTimeStampRaw

FileWriteOption

OperationId

QueryId