nx

Module fs

FileSystem support

Re-exports

pub use fsp::fsp_sf::DirectoryEntry;

pub use fsp::fsp_sf::DirectoryEntryType;

pub use fsp::fsp_sf::DirectoryOpenMode;

pub use fsp::fsp_sf::FileAttribute;

pub use fsp::fsp_sf::FileOpenMode;

pub use fsp::fsp_sf::FileQueryRangeInfo;

pub use fsp::fsp_sf::FileReadOption;

pub use fsp::fsp_sf::FileTimeStampRaw;

pub use fsp::fsp_sf::FileWriteOption;

pub use fsp::fsp_sf::OperationId;

pub use fsp::fsp_sf::QueryId;

Modules

rc

FS-related result definitions

subdir

Helper object types for treating FS subdirectories as IPC filesystems

Structs

DirectoryAccessor

Represents a wrapper type to simplify directory access

FileAccessor

Represents a wrapper type to simplify file access, tracking the currently seek-ed location in the file.

FileOpenOption

Represents options for opening files

ProxyDirectory

Represents a wrapper Directory implementation to translate IPC IDirectoryClient objects to Directory objects.

ProxyFile

Represents a wrapper File implementation to translate IPC IFileClient objects to File objects.

ProxyFileSystem

Represents a wrapper FileSystem implementation to translate IPC IFileSystemClient objects to FileSystem objects

Enums

SeekFrom

Represents an offset kind/relativeness.

Traits

Directory

Represents a directory.

File

Represents a file, abstracted from the IPC client API.

FileSystem

Represents a filesystem.

Functions

commit

Commits on a filesystem

convert_file_open_mode_to_option

Converts a FileOpenMode to a FileOpenOption

convert_file_open_option_to_mode

Converts a FileOpenOption to a FileOpenMode

create_directory

Creates a directory

create_file

Creates a file

format_path

Returns the FileSystem and the processed path corresponding to a given path

get_entry_type

Gets a path’s DirectoryEntryType

get_file_time_stamp_raw

Gets the FileTimeStampRaw of a file

get_free_space_size

Gets the free space size at a given path

get_fspsrv_session

Gets the global IFileSystemProxyClient shared object used for fsp-srv support

get_path_filesystem

Returns the FileSystem corresponding to a given path

get_total_space_size

Gets the total space size at a given path

initialize_fspsrv_session

Initializes fsp-srv support instantiating a FileSystemProxy shared object

is_fspsrv_session_initialized

Gets whether fsp-srv support was initialized

mount

Mounts a FileSystem

mount_fsp_filesystem

Mounts an IPC IFileSystemClient

mount_sd_card

Mounts the system’s SD card using fsp-srv support

open_directory

Opens a directory as a DirectoryAccessor

open_file

Opens a file as a FileAccessor

query_entry

Queries on a path

remove_children_all

Deletes all the children files/directories inside a directory

remove_dir

Deletes a directory

remove_dir_all

Deletes a directory and all its children files/directories

remove_file

Deletes a file

rename

Renames a file/directory

rename_directory

Renames a directory

rename_file

Renames a file

set_concatenation_file_attribute

Sets the “concatenation file” attribute on a file.

unmount

Unmounts a mounted filesystem

unmount_all

Unmounts all filesystems