nx::fs

Trait Directory

pub trait Directory: Sync {
    // Required methods
    fn read(&self, out_entries: &mut [DirectoryEntry]) -> Result<usize>;
    fn get_entry_count(&self) -> Result<u64>;
}

Represents a directory.

Required Methods

fn read(&self, out_entries: &mut [DirectoryEntry]) -> Result<usize>

Reads existing entries, returning the actual number of read entries.

The max number of entries to read is determined by the output slice size and the actually existing entry count.

Arguments:

out_entries: The out DirectoryEntry slice to fill.

fn get_entry_count(&self) -> Result<u64>

Gets the Directory’s entry count.