Trait nx::fs::FileSystem

pub trait FileSystem {

    // Required methods
    fn create_file(
        &mut self,
        path: String,
        attribute: FileAttribute,
        size: usize
    ) -> Result<()>;
    fn delete_file(&mut self, path: String) -> Result<()>;
    fn create_directory(&mut self, path: String) -> Result<()>;
    fn delete_directory(&mut self, path: String) -> Result<()>;
    fn delete_directory_recursively(&mut self, path: String) -> Result<()>;
    fn rename_file(&mut self, old_path: String, new_path: String) -> Result<()>;
    fn rename_directory(
        &mut self,
        old_path: String,
        new_path: String
    ) -> Result<()>;
    fn get_entry_type(&mut self, path: String) -> Result<DirectoryEntryType>;
    fn open_file(
        &mut self,
        path: String,
        mode: FileOpenMode
    ) -> Result<Shared<dyn File>>;
    fn open_directory(
        &mut self,
        path: String,
        mode: DirectoryOpenMode
    ) -> Result<Shared<dyn Directory>>;
    fn commit(&mut self) -> Result<()>;
    fn get_free_space_size(&mut self, path: String) -> Result<usize>;
    fn get_total_space_size(&mut self, path: String) -> Result<usize>;
    fn clean_directory_recursively(&mut self, path: String) -> Result<()>;
    fn get_file_time_stamp_raw(
        &mut self,
        path: String
    ) -> Result<FileTimeStampRaw>;
    fn query_entry(
        &mut self,
        path: String,
        query_id: QueryId,
        in_buf: *const u8,
        in_buf_size: usize,
        out_buf: *mut u8,
        out_buf_size: usize
    ) -> Result<()>;

}

Represents a filesystem

Required Methods

fn create_file( &mut self, path: String, attribute: FileAttribute, size: usize ) -> Result<()>

Creates a file

Arguments

path: The path to use

size: The initial file size

attribute: The file attribute

fn delete_file(&mut self, path: String) -> Result<()>

Deletes a file

Arguments

path: The path to use

fn create_directory(&mut self, path: String) -> Result<()>

Creates a directory

Arguments

path: The path to use

fn delete_directory(&mut self, path: String) -> Result<()>

Deletes a directory

Arguments

path: The path to use

fn delete_directory_recursively(&mut self, path: String) -> Result<()>

Deletes a directory and all its children files/directories

Arguments

path: The path to use

fn rename_file(&mut self, old_path: String, new_path: String) -> Result<()>

Renames a file

Arguments

old_path: The old path to use

new_path: The new path to use

fn rename_directory(&mut self, old_path: String, new_path: String) -> Result<()>

Renames a directory

Arguments

old_path: The old path to use

new_path: The new path to use

fn get_entry_type(&mut self, path: String) -> Result<DirectoryEntryType>

Gets a path’s DirectoryEntryType

Arguments

path: The path to use

fn open_file( &mut self, path: String, mode: FileOpenMode ) -> Result<Shared<dyn File>>

Opens a File

Arguments

path: The path to use

mode: The open mode

fn open_directory( &mut self, path: String, mode: DirectoryOpenMode ) -> Result<Shared<dyn Directory>>

Opens a Directory

Arguments

path: The path to use

mode: The open mode

fn commit(&mut self) -> Result<()>

Commits the filesystem

fn get_free_space_size(&mut self, path: String) -> Result<usize>

Gets the free space size at a given path

Argument

path: The path to use

fn get_total_space_size(&mut self, path: String) -> Result<usize>

Gets the total space size at a given path

Argument

path: The path to use

fn clean_directory_recursively(&mut self, path: String) -> Result<()>

Deletes all the children files/directories inside a directory

Arguments

path: The path to use

fn get_file_time_stamp_raw(&mut self, path: String) -> Result<FileTimeStampRaw>

Gets the FileTimeStampRaw of a file

Arguments

path: The path to use

fn query_entry( &mut self, path: String, query_id: QueryId, in_buf: *const u8, in_buf_size: usize, out_buf: *mut u8, out_buf_size: usize ) -> Result<()>

Queries on a path

Arguments

query_id: The QueryId

in_buf: Input data

in_buf_size: Input data size

out_buf: Output data

out_buf_size: Output data size