Struct nx::gpu::Context

pub struct Context { /* private fields */ }

Represents a graphics context

Implementations

impl Context

pub fn new( nv_kind: NvDrvServiceKind, vi_kind: ViServiceKind, transfer_mem_size: usize ) -> Result<Self>

Creates a new Context

This automatically accesses VI and NV INvDrvServices services (of the specified kinds) and creates NV transfer memory

Arguments

nv_kind: The NvDrvServiceKind

vi_kind: The ViServiceKind

transfer_mem_size: The transfer memory size to use

pub fn from( vi_srv: Shared<dyn IObject>, application_display_srv: Shared<dyn IApplicationDisplayService>, nvdrv_srv: Shared<dyn INvDrvServices>, transfer_mem_size: usize ) -> Result<Self>

Creates a new Context with already existing service objects

This automatically creates NV transfer memory

Arguments

vi_srv: The VI service object

application_display_srv: The VI IApplicationDisplayService interface object

nvdrv_srv: The NV INvDrvServices service object

transfer_mem_size: The transfer memory size to use

pub fn get_nvdrv_service(&self) -> Shared<dyn INvDrvServices>

Gets the underlying NV INvDrvServices service object

pub fn get_application_display_service( &self ) -> Shared<dyn IApplicationDisplayService>

Gets the underlying IApplicationDisplayService object

pub fn get_hos_binder_driver(&self) -> Shared<dyn IHOSBinderDriver>

Gets the underlying IHOSBinderDriver object

pub fn create_stray_layer_surface( &mut self, display_name: &str, buffer_count: u32, color_fmt: ColorFormat, pixel_fmt: PixelFormat, layout: Layout ) -> Result<Surface>

Creates a Surface based on a stray layer

Arguments

display_name: The display where the stray layer is created

buffer_count: The buffer count to use internally

color_fmt: The color format to use

pixel_fmt: The pixel format to use

layout: The layout type to use

pub fn create_managed_layer_surface( &mut self, display_name: &str, aruid: AppletResourceUserId, layer_flags: LayerFlags, x: f32, y: f32, width: u32, height: u32, z: LayerZ, buffer_count: u32, color_fmt: ColorFormat, pixel_fmt: PixelFormat, layout: Layout ) -> Result<Surface>

Creates a Surface based on a managed layer

This is a more privileged layer than stray layers

Arguments

display_name: The display where the stray layer is created

aruid: The AppletResourceUserId value to use, can be zero

layer_flags: The LayerFlags value to use

x: The layer X position

y: The layer Y position

width: The layer width

height: The layer height

z: The layer Z position (depth)

buffer_count: The buffer count to use internally

color_fmt: The color format to use

pixel_fmt: The pixel format to use

layout: The layout type to use