nx::gpu

Struct Context

pub struct Context { /* private fields */ }

Represents a graphics context

Implementations

impl Context

pub fn new( nv_kind: NvDrvServiceKind, vi_kind: ViServiceKind, transfer_mem_size: usize, ) -> Result<Self>

Creates a new Context

This automatically accesses VI and NV INvDrvClient services (of the specified kinds) and creates NV transfer memory

Arguments

nv_kind: The NvDrvServiceKind

vi_kind: The ViServiceKind

transfer_mem_size: The transfer memory size to use

pub fn from( vi_srv: RootServiceHolder, application_display_srv: Box<dyn IApplicationDisplayClient>, nvdrv_srv: Box<dyn INvDrvClient>, transfer_mem_size: usize, nv_host_as_gpu: bool, ) -> Result<Self>

Creates a new Context with already existing service objects

This automatically creates NV transfer memory

Arguments

vi_srv: The VI service object

application_display_srv: The vi IApplicationDisplayClient interface object

nvdrv_srv: The NV INvDrvClient service object

transfer_mem_size: The transfer memory size to use

nv_host_as_gpu: Flag whether to open a handle to the GPU for hardware accelerated rendering.

pub fn nvdrv_service(&self) -> &dyn INvDrvClient

Gets the underlying NV INvDrvClient service object

pub fn nvdrv_service_mut(&mut self) -> &mut dyn INvDrvClient

Gets the underlying NV INvDrvClient service object mutably

pub fn get_application_display_service(&self) -> &dyn IApplicationDisplayClient

Gets the underlying IApplicationDisplayClient object

pub fn get_application_display_service_mut( &mut self, ) -> &mut dyn IApplicationDisplayClient

Gets the underlying IApplicationDisplayClient object mutably

pub fn get_hos_binder_driver(&self) -> Arc<dyn IHOSBinderDriverClient>

Gets the underlying IHOSBinderDriverClient object