nx::gpu::surface

Struct Surface

pub struct Surface { /* private fields */ }

Represents a wrapper around layer manipulation

Implementations

impl Surface

pub fn new_stray( gpu_ctx: Arc<RwLock<Context>>, display_name: &str, buffer_count: u32, block_config: BlockLinearHeights, color_fmt: ColorFormat, pixel_fmt: PixelFormat, ) -> Result<Self>

Create a new stray layer covering the full screen (modeled as a 1280x720 pixel surface)

Arguments:

gpu_ctx: a handle to a gpu shared context we can use to control the propeties of our layer.

display_name: a name for the spawned layer

buffer_count: the number of buffers to use. using a value of 0 will error, and a value of 1 will hang at runtime as you cannot dequeue the active buffer.

block_config: configuration value for the GPU buffer tiling height. Applications with row based rendering should set a low value and applications that draw mainly in blocks (e.g. image decoding or text) should set medium or higher values.

color_format: The color format set on the layer

pixel_format: The pixel format for the buffer. Must match the color format to run without graphical issues.

pub fn new_managed( gpu_ctx: Arc<RwLock<Context>>, display_name: &str, aruid: AppletResourceUserId, layer_flags: LayerFlags, x: f32, y: f32, z: LayerZ, framebuffer_width: u32, framebuffer_height: u32, buffer_count: u32, block_config: BlockLinearHeights, color_fmt: ColorFormat, pixel_fmt: PixelFormat, scaling: ScaleMode, ) -> Result<Self>

Create a new stray layer covering the full screen (modeled as a 1280x720 pixel surface)

Arguments:

gpu_ctx: a handle to a gpu shared context we can use to control the propeties of our layer.

display_name: a name for the spawned layer

aruid: the applet resource user ID for the running app (usually 0, but can be retrieved by initializing the applet module).

layer_flags: the vi service flags to set for the layer (usually 0).

x: the offset of the layer from the left edge of the screen.

y: the offset of the layer from the top edge of the screen.

width: width of the canvas.

height: height of the canvas.

buffer_count: the number of buffers to use. using a value of 0 will error, and a value of 1 will hang at runtime as you cannot dequeue the active buffer.

block_config: configuration value for the GPU buffer tiling height. Applications with row based rendering should set a low value and applications that draw mainly in blocks (e.g. image decoding or text) should set medium or higher values.

color_format: The color format set on the layer

pixel_format: The pixel format for the buffer. Must match the color format to run without graphical issues.

scaling: The configuration for mapping the framebuffer/canvas onto the spawned layer which may be a larger/smaller size.

pub fn new_from_parts( binder_handle: BinderHandle, gpu_ctx: Arc<RwLock<Context>>, buffer_count: u32, display_id: DisplayId, layer_id: LayerId, width: u32, height: u32, block_height_config: BlockLinearHeights, color_fmt: ColorFormat, pixel_fmt: PixelFormat, managed: bool, ) -> Result<Self>

Creates a new Surface

This is not meant to really be used manually, see Context

pub fn get_block_linear_config(&self) -> BlockLinearHeights

pub fn dequeue_buffer( &mut self, is_async: bool, ) -> Result<(*mut u8, usize, i32, bool, MultiFence)>

Dequeues a buffer, returning the buffer address, its size, its slot, whether it has fences, and those mentioned fences

Arguments

is_async: Whether to dequeue asynchronously

pub fn queue_buffer(&mut self, slot: i32, fences: MultiFence) -> Result<()>

Queues a buffer

Arguments

slot: The buffer slot

fences: The buffer fences

pub fn wait_fences(&mut self, fences: MultiFence, timeout: i32) -> Result<()>

Waits for the given fences

Arguments

fences: The fences

timeout: The wait timeout

pub fn set_visible(&mut self, visible: bool) -> Result<()>

Sets whether the surface (its layer) is visible

Arguments

visible: Whether its visible

pub fn push_layer_stack(&mut self, stack_type_id: LayerStackId) -> Result<()>

Adds a layer visibility stack

The Management display server maintains a stack of visible layers. This allows us to order our layer relative to other elements (e.g. visible or invisible to screenshots/recordings)

Atguments

stack_type_id: the type of layer to add to the visibility stack

pub fn wait_buffer_event(&self, timeout: i64) -> Result<()>

Waits for the buffer event

Arguments

timeout: The wait timeout

pub fn wait_vsync_event(&self, timeout: i64) -> Result<()>

Waits for the vsync event

Arguments

timeout: The wait timeout

pub const fn width(&self) -> u32

Gets the surface width

pub const fn height(&self) -> u32

Gets the surface height

pub const fn color_format(&self) -> ColorFormat

Gets the surface ColorFormat

pub const fn stride(&self) -> u32

Computes and gets the surface stride (distance between adjacent rows in pixels, incliuding padding).

pub const fn pitch(&self) -> u32

pub const fn single_buffer_size(&self) -> u32