nx::gpu::surface
Struct Surface

pub struct Surface { /* private fields */ }

Represents a wrapper around layer manipulation

Implementations

impl Surface

pub fn new( binder_handle: BinderHandle, nvdrv_srv: Shared<dyn INvDrvServices>, application_display_service: Shared<ApplicationDisplayService>, nvhost_fd: u32, nvmap_fd: u32, nvhostctrl_fd: u32, hos_binder_driver: Shared<dyn IHOSBinderDriver>, buffer_count: u32, display_id: DisplayId, layer_id: LayerId, width: u32, height: u32, color_fmt: ColorFormat, pixel_fmt: PixelFormat, layout: Layout, layer_destroy_fn: LayerDestroyFn, ) -> Result<Self>

Creates a new Surface

This is not meant to really be used manually, see Context

Arguments

binder_handle: The binder handle to use

nvdrv_srv: The NvDrvServicesService dyn-object to use

``

pub fn dequeue_buffer( &mut self, is_async: bool, ) -> Result<(*mut u8, usize, i32, bool, MultiFence)>

Dequeues a buffer, returning the buffer address, its size, its slot, whether it has fences, and those mentioned fences

Arguments

is_async: Whether to dequeue asynchronously

pub fn queue_buffer(&mut self, slot: i32, fences: MultiFence) -> Result<()>

Queues a buffer

Arguments

slot: The buffer slot

fences: The buffer fences

pub fn wait_fences(&mut self, fences: MultiFence, timeout: i32) -> Result<()>

Waits for the given fences

Arguments

fences: The fences

timeout: The wait timeout

pub fn set_visible(&mut self, visible: bool) -> Result<()>

Sets whether the surface (its layer) is visible

Arguments

visible: Whether its visible

pub fn push_layer_stack(&mut self, stack_type_id: LayerStackId) -> Result<()>

Adds a layer visibility stack

The Management display server maintains a stack of visible layers. This allows us to order our layer relative to other elements (e.g. visible or invisible to screenshots/recordings)

Atguments

stack_type_id: the type of layer to add to the visibility stack

pub fn wait_buffer_event(&mut self, timeout: i64) -> Result<()>

Waits for the buffer event

Arguments

timeout: The wait timeout

pub fn wait_vsync_event(&mut self, timeout: i64) -> Result<()>

Waits for the vsync event

Arguments

timeout: The wait timeout

pub fn get_width(&self) -> u32

Gets the surface width

pub fn get_height(&self) -> u32

Gets the surface height

pub fn get_color_format(&self) -> ColorFormat

Gets the surface ColorFormat

pub fn compute_stride(&self) -> u32

Computes and gets the surface stride