nx

Module hbl

HBL (homebrew loader) ABI support and utils

Structs

AbiConfigAppletFlags

Represents optional flag values for the specific case of AbiConfigEntryKey::AppletType config entries

AbiConfigEntry

Represents an ABI config entry layout

AbiConfigEntryFlags

Represents optional flags for config entries

AtomicAppletType

A wrapper around AppletType which can be safely shared between threads.

Version

Represents the hbl-ABI format of the system version

Enums

AbiConfigEntryKey

Represents the entry value keys for a hbl ABI context

AppletType

Represents the applet types for HBL

Statics

G_NEXT_LOAD_ARGV

G_NEXT_LOAD_PATH

Functions

get_applet_type

Gets the current applet type (according to HBL)

get_last_load_result

Gets the last load ResultCode

get_loader_info

Gets the loader information string, about HBL

get_next_load_argv

Gets the next load argv, AKA the argv of the homebrew NRO which will be executed after this one exits

get_next_load_path

Gets the next load path, AKA the path of the homebrew NRO which will be executed after this one exits

get_process_handle

Gets the current process handle

get_random_seed

Gets the random seed values sent by HBL

set_next_load_entry

Sets the next homebrew NRO (path and argv) to execute after this one exits