nx::input
Struct Context

pub struct Context { /* private fields */ }

Represents a simple type for dealing with input handling

Implementations

impl Context

pub fn new( supported_style_tags: NpadStyleTag, player_count: usize, ) -> Result<Self>

Creates a Context from supported input values

The supported values are essentially used to enable supported controllers/controller configs via hid commands, and for opening Player types

Arguments

supported_style_tags: Supported NpadStyleTag flags

supported_npad_ids: Supported NpadIdType values

pub fn get_player(&self, npad_id: NpadIdType) -> Player

Opens a Player type for the specified NpadIdType

This simplifies creating a Player type, since this context contains the supported NpadStyleTag values and the mapped shared-memory address

Arguments

npad_id: The NpadIdType to use

pub fn get_touches(&mut self, touch_states: &mut [TouchState]) -> usize

Gets the current TouchState values for the console touch-screen, returning the number of states present/set

Arguments

touch_states: Array of TouchState values to get filled, the array doesn’t have to be bigger than 17 items