nx::input

Struct Player

pub struct Player<'player> { /* private fields */ }

Represents a console controller type

It’s essentially a wrapper type over HID shared-memory to simplify input detection

Implementations

impl<'player, 'context: 'player> Player<'player>

pub fn new( npad_id: NpadIdType, supported_style_tags: NpadStyleTag, shmem: &'context SharedMemoryFormat, ) -> Result<Self>

Creates a Player from shared-memory information

If using a Context, look for Context::get_player instead (for simplicity)

Arguments

npad_id: The NpadIdType of the desired controller

supported_style_tags: The NpadStyleTag flags which will be used by the Player to scan for input, etc.

shmem_ptr: The address of HID shared-memory

pub fn get_previous_buttons(&self) -> NpadButton

pub fn get_style_tag_attributes(&self, style_tag: NpadStyleTag) -> NpadAttribute

Gets the NpadAttributes for a certain NpadStyleTag

Arguments

style_tag: Must be a NpadStyleTag with a single flag set (otherwise only one will take effect and the rest will be ignored)

pub fn get_stick_status( &self, style_tag: NpadStyleTag, ) -> (AnalogStickState, AnalogStickState)

Gets the stick status from a provided style tag (which may or may not be configured)

pub fn get_first_stick_status( &self, deadzone: f32, ) -> Option<(AnalogStickState, AnalogStickState)>

pub fn get_style_tag_buttons(&mut self, style_tag: NpadStyleTag) -> NpadButton

Gets the NpadButtons for a certain NpadStyleTag

Arguments

style_tag: Must be a NpadStyleTag with a single flag set (otherwise only one will take effect and the rest will be ignored)

pub fn get_style_tag_buttons_down( &mut self, style_tag: NpadStyleTag, ) -> NpadButton

Gets the down NpadButtons for a certain NpadStyleTag

This is similar to get_style_tag_buttons but this only gets the buttons once after they’re down/pressed

Arguments

style_tag: Must be a NpadStyleTag with a single flag set (otherwise only one will take effect and the rest will be ignored)

pub fn get_style_tag_buttons_up( &mut self, style_tag: NpadStyleTag, ) -> NpadButton

Gets the up NpadButtons for a certain NpadStyleTag

This is similar to get_style_tag_buttons but this only gets the buttons once after they’re up/released

Arguments

style_tag: Must be a NpadStyleTag with a single flag set (otherwise only one will take effect and the rest will be ignored)

pub fn get_buttons(&mut self) -> NpadButton

Gets the NpadButtons for all of the supported NpadStyleTags, combining all of them

This is like combining the result of get_style_tag_buttons with all the supported NpadStyleTag flags

pub fn get_buttons_down(&mut self) -> NpadButton

Gets the down NpadButtons for all of the supported NpadStyleTags, combining all of them

This is similar to get_buttons but this only gets the buttons once after they’re down/pressed

pub fn get_buttons_up(&mut self) -> NpadButton

Gets the up NpadButtons for all of the supported NpadStyleTags, combining all of them

This is similar to get_buttons but this only gets the buttons once after they’re up/released

pub fn get_button_updates( &mut self, ) -> (NpadButton, NpadButton, NpadButton, NpadButton)

Gets the up NpadButtons for all of the supported NpadStyleTags, combining all of them

This only updates the state once, but is otherwise equivalent to (self.get_previous(), self.get_buttons(), self.get_buttons_down(), self.get_buttons_up())

pub fn get_npad_id(&self) -> NpadIdType

Gets the NpadIdType being used with this Player

pub fn get_supported_style_tags(&self) -> NpadStyleTag

Gets the supported NpadStyleTag flags being used with this Player

pub fn get_reported_style_tag(&self) -> NpadStyleTag

pub fn get_controller_type(&self) -> DeviceType