nx::ipc::sf

Module applet

Re-exports

pub use super::AppletResourceUserId;

Structs

AllSystemAppletProxiesService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

AppletAttribute

AppletProcessLaunchReason

ApplicationProxy

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

AudioController

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

CommonStateGetter

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

DisplayController

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

LibraryAppletAccessor

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

LibraryAppletCreator

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

LibraryAppletProxy

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

OverlayAppletProxy

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

ProcessWindingController

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SelfController

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

Storage

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

StorageAccessor

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SystemAppletProxy

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SystemApplicationProxy

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

WindowController

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

Enums

AppletId

AppletMessage

LibraryAppletMode

OperationMode

PerformanceMode

ScreenShotPermission

Traits

IAllSystemAppletProxiesServiceClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IAllSystemAppletProxiesServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IApplicationProxyClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IApplicationProxyServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IAudioControllerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IAudioControllerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ICommonStateGetterClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ICommonStateGetterServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IDisplayControllerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IDisplayControllerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ILibraryAppletAccessorClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILibraryAppletAccessorServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ILibraryAppletCreatorClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILibraryAppletCreatorServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ILibraryAppletProxyClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ILibraryAppletProxyServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IOverlayAppletProxyClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IOverlayAppletProxyServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IProcessWindingControllerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IProcessWindingControllerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISelfControllerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISelfControllerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IStorageAccessorClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IStorageAccessorServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IStorageClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IStorageServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISystemAppletProxyClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISystemAppletProxyServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISystemApplicationProxyClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISystemApplicationProxyServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IWindowControllerClient

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IWindowControllerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ProxyCommon