nx::ipc::sf

Module hid

Modules

shmem

Structs

AnalogStickCalibrationFlags

AnalogStickState

AppletFooterUiAttribute

AppletResource

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

BasicXpadAttribute

BasicXpadButton

CaptureButton

DebugPadAttribute

DebugPadButton

DeviceType

DigitizerAttribute

DigitizerButton

DirectionState

GestureAttribute

GesturePoint

HidServer

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

HidSysServer

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

HomeButton

InputSourceState

KeyboardKey

KeyboardModifier

LockKeyFlags

MouseAttribute

MouseButton

NpadAttribute

NpadButton

NpadControllerColor

NpadStyleTag

NpadSystemButtonProperties

NpadSystemProperties

SixAxisSensorAttribute

SixAxisSensorProperties

SixAxisSensorUserCalibrationFlags

SleepButton

TouchAttribute

TouchState

Enums

AnalogStickManualCalibrationStage

AppletFooterUiType

ColorAttribute

GestureDirection

GestureType

NpadIdType

NpadJoyAssignmentMode

NpadJoyDeviceType

NpadLagerType

NpadLarkType

NpadLuciaType

SixAxisSensorUserCalibrationStage

UniquePadInterface

UniquePadType

Traits

IAppletResource

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IAppletResourceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IHidServer

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IHidServerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IHidSysServer

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IHidSysServerServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Type Aliases

NpadBatteryLevel

UniquePadSerialNumber