nx::ipc::sf
Module vi

Structs

ApplicationDisplayService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

ApplicationRootService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

LayerFlags

ManagerDisplayService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

ManagerRootService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SystemDisplayService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

SystemRootService

The default client for the $t trait. All implementors of the trait need to read their session in accordance with this Types IPC Parameter traits.

Enums

DisplayServiceMode

LayerStackId

Traits

IApplicationDisplayService

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IApplicationDisplayServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IApplicationRootService

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IApplicationRootServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IManagerDisplayService

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IManagerDisplayServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

IManagerRootService

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

IManagerRootServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISystemDisplayService

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISystemDisplayServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

ISystemRootService

The IPC server trait for the [<I $intf Server>] trait. All methods are provided and overriding the implementations is not recommended.

ISystemRootServiceServer

The IPC server trait for the [<I $intf Server>] trait. The methods not provided must all be implemented, but overriding the IPC wrappers is not recommended.

Type Aliases

DisplayId

DisplayName

LayerId