nx
Module la

Library applet support and utils

Structs

CommonArguments

Represents the common arguments layout sent as starting input by/to all library applets

LibraryAppletHolder

Represents a wrapper type for using library applets

Functions

create_library_applet

Creates a LibraryAppletHolder from the given library applet params

create_write_storage

Wrapper for creating a IStorage shared object from the given value

get_creator

Gets access to the global ILibraryAppletCreator shared object instance

initialize

Initializes library applet support with the provided ILibraryAppletCreator shared object

is_initialized

Gets whether library applet support was initialized

launch_wait_library_applet

Wrapper to create, launch and wait for a library applet, expecting simple input and output data

read_storage

Wrapper for reading data from a IStorage shared object

write_storage

Wrapper for writing data to a IStorage shared object