nx::la

Struct LibraryAppletHolder

pub struct LibraryAppletHolder { /* private fields */ }

Represents a wrapper type for using library applets

Implementations

impl LibraryAppletHolder

pub fn new(accessor: Box<dyn ILibraryAppletAccessorClient>) -> Result<Self>

Creates a LibraryAppletHolder from an existing ILibraryAppletAccessorClient shared object

This shouldn’t be manually created unless the accessor object was obtained manually (see create_library_applet)

pub fn get_accessor(&self) -> &dyn ILibraryAppletAccessorClient

Gets a reference to the underlying ILibraryAppletAccessorClient shared object

pub fn get_accessor_mut(&mut self) -> &dyn ILibraryAppletAccessorClient

Gets a mutable reference to the underlying ILibraryAppletAccessorClient shared object

pub fn push_in_data_storage(&mut self, storage: Storage) -> Result<()>

Pushes an input IStorageClient shared object to the library applet

pub fn push_in_data<T: Copy>(&mut self, t: T) -> Result<()>

Pushes input data to the library applet

This is a wrapper which creates an IStorageClient object with the given value and pushes it

pub fn start(&mut self) -> Result<()>

Starts the library applet

pub fn join(&mut self, timeout: Option<i64>) -> Result<()>

Waits until the library applet’s state-changed event signals

This effectively waits until the library applet exits or the timeout expires

pub fn pop_out_data_storage(&mut self) -> Result<Storage>

Pops an output IStorageClient shared object from the library applet

pub fn pop_out_data<O: Copy>(&mut self) -> Result<O>

Pops output data from the library applet

This is a wrapper which pops an IStorageClient object and reads its data (reads [size_of][cmem::size_of] O bytes and returns that data)