Macro nx::ipc_sf_object_impl_default_command_metadata

macro_rules! ipc_sf_object_impl_default_command_metadata {
    () => { ... };
}

Macro to simplify defining an IPC interface impl command metadata

This is meant to only be used inside IObject impls!

Note that this only has to be manually used for non-client-IPC interface types - for client-IPC interface types, see ipc_client_define_object_default

Example

use nx::ipc::sf::{Session, IObject};
 
// Let's assume an IPC interface named "IExampleInterface" exists
// Let's then create a custom implementation of that interface
pub struct ExampleInterface {
    // Required, only effectively used on client IPC interfaces
    dummy_session: Session
}
 
impl IObject for ExampleInterface {
    ipc_sf_object_impl_default_command_metadata!();
 
    fn get_session(&mut self) -> &mut Session {
        &mut self.dummy_session
    }
}
 
impl IExampleInterface for ExampleInterface {
    (...)
}