nx::mem

Module alloc

Allocator implementation and definitions

Modules

rc

Allocation-specific result definitions

Structs

Buffer

Represents a wrapped and manually managed heap value

Layout

Layout of a block of memory.

Constants

PAGE_ALIGNMENT

Statics

HEAP_SIZE

Traits

AllocatorEx

Represents a heap allocator for this library

Functions

configure_heap

Default implementation

initialize

Initializes the global allocator with the given address and size. Returns a bool to indicate if the memory was consumed by the allocator

is_enabled

Gets whether heap allocations are enabled