nx

Module service

Base service/named port support and wrappers

Modules

applet

“appletAE” service definitions.

dispdrv

“dispdrv” service definitions.

fatal

“fatal:u” service definitions.

fsp

“fsp-srv” service definitions.

hid

“hid” service definitions.

ldr

“ldr:shel” service definitions.

lm

“lm” service definitions.

lr

“lr” service definitions.

mii

“mii:e” service definitions.

ncm

“ncm” service definitions.

nfp

“nfp:*” service definitions.

nv

“nvdrv” and “nvdrv:*” service definitions.

pm

“pm:*” service definitions.

psc

“psc:m” service definitions.

psm

“psm” service definitions.

set

“set:sys” service definitions.

sm

spl

“csrng” service definitions.

usb

“usb:hs” service definitions.

vi

“vi:*” service definitions.

Traits

INamedPort

Represents a named port interface

IService

Represents a service interface

Functions

new_named_port_object

Wrapper for connecting to a named port and instantiating the wrapper interface over the specified named port

new_service_object

Wrapper for accessing a service and instantiating the wrapper interface over the specified service