nx::service

Trait IService

pub trait IService: IClientObject {
    // Required methods
    fn get_name() -> ServiceName;
    fn as_domain() -> bool;
    fn post_initialize(&mut self) -> Result<()>;
}

Represents a service interface

Interfaces which wrap services (see SM) must implement this trait

Required Methods

fn get_name() -> ServiceName

Gets the service’s name

fn as_domain() -> bool

Gets whether the service should be used as a domain

If this is [true], the service will be converted to a domain after being accessed (see convert_to_domain) in new_service_object

fn post_initialize(&mut self) -> Result<()>

This will get executed after accessing the service in new_service_object, allowing for extra initialization

Some interfaces may have initialization commands (check SM’s case) which can be automatically called this way

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.