nx::smc

Function generate_random_bytes

pub fn generate_random_bytes(out_bytes: &mut [u8]) -> Result

Secure monitor call which generates random bytes

Note that the process needs to be running in processor 3 in order to be able to execute secure monitor calls

Arguments

out_bytes: Array to fill with random bytes, whose size mustn’t exceed GENERATE_RANDOM_BYTES_MAX_SIZE