Struct nx::thread::Thread

#[repr(C)]
pub struct Thread {

    pub self_ref: *mut Thread,
    pub state: ThreadState,
    pub owns_stack: bool,
    pub pad: [u8; 2],
    pub handle: Handle,
    pub stack: *mut u8,
    pub stack_size: usize,
    pub reserved: [u8; 38],
    pub version: u16,
    pub reserved_2: [u8; 248],
    pub entry: Option<ThreadEntry>,
    pub reserved_3: [u8; 40],
    pub name: ThreadName,
    pub name_addr: *mut u8,
    pub reserved_4: [u8; 32],

}

Represents a thread

Fields
self_ref: *mut Threadstate: ThreadStateowns_stack: boolpad: [u8; 2]handle: Handlestack: *mut u8stack_size: usizereserved: [u8; 38]version: u16reserved_2: [u8; 248]entry: Option<ThreadEntry>reserved_3: [u8; 40]name: ThreadNamename_addr: *mut u8reserved_4: [u8; 32]
Implementations

impl Thread

pub const fn empty() -> Self

Creates an empty, thus invalid Thread

pub fn new_remote( handle: Handle, name: &str, stack: *mut u8, stack_size: usize ) -> Result<Self>

Creates a Thread from an existing thread handle, a name, and stack

Arguments

handle: The remote thread handle

name: The custom thread name

stack: The remote stack address

stack_size: The remote stack size

pub fn new_with_stack<T: Copy, F: 'static + Fn(&T)>( entry: F, args: &T, name: &str, stack: *mut u8, stack_size: usize ) -> Result<Self>

Creates a new Thread with an entrypoint + args, name and stack

Note that it needs to be initialized (Thread::initialize) before being started (Thread::start)

Arguments

entry: The entrypoint function, taking args

args: The entrypoint arguments

name: The desired thread name

stack: The stack address

stack_size: The stack size

pub fn new<T: Copy, F: 'static + Fn(&T)>( entry: F, args: &T, name: &str, stack_size: usize ) -> Result<Self>

Creates a new Thread with an entrypoint + args, name and stack

Same as calling Thread::new_with_stack but with the stack being automatically allocated from heap

Note that it needs to be initialized (Thread::initialize()) before being started (Thread::start())

Arguments

entry: The entrypoint function, taking args

args: The entrypoint arguments

name: The desired thread name

stack_size: The desired stack size

pub fn initialize(&mut self, priority: i32, processor_id: i32) -> Result<()>

Initializes a Thread

Technically, this actually “creates” it using svc::create_thread

It must be in ThreadState::NotInitialized, otherwise this will fail with ResultInvalidState

Arguments

priority: The desired priority, or PRIORITY_AUTO to use the current thread value

processor_id: The desired processor ID, in [0-3] range or svc::DEFAULT_PROCESS_PROCESSOR_ID for the process’s default value

pub fn start(&mut self) -> Result<()>

Starts the Thread

It must be in ThreadState::Initialized or ThreadState::Terminated, otherwise this will fail with ResultInvalidState

Essentially uses svc::start_thread

pub fn join(&mut self) -> Result<()>

Joins the Thread

It must be in ThreadState::Started, otherwise this will fail with ResultInvalidState

Essentially waits for the thread handle, which will signal when it finishes

pub fn is_remote(&self) -> bool

Gets whether this Thread is remote

pub fn get_handle(&self) -> Handle

Gets this Thread’s handle

pub fn get_priority(&self) -> Result<i32>

Gets this Thread’s priority

It must be in any state but ThreadState::NotInitialized, otherwise this will fail with ResultInvalidState

Essentially uses svc::get_thread_priority

pub fn get_id(&self) -> Result<u64>

Gets this Thread’s ID

It must be in any state but ThreadState::NotInitialized, otherwise this will fail with ResultInvalidState

Essentially uses svc::get_thread_id