Module nx::util

Common library utilities

Modules

rc

Util-specific result definitions

Structs

CString

Represents a C-like string of a given size (mostly like a C char[S] array)

CString16

Represents a C-like 16-bit string of a given size (mostly like a C char16_t[S] array)

PointerAndSize

Represents a pair of a pointer and a size

Uuid

Represents a 16-byte UUID

Functions

raw_transmute⚠

Works like core::mem::transmute, but being less restrictive and extremely more unsafe

simple_panic_handler

Simplified panic handler using a provided Logger type, available as a helpful default panic handler

str_copy

Copies one &str into another, and returns the destination &str

str_ptr_len

Same as C’s strlen()