nx

Module util

Common library utilities

Modules

rc

Util-specific result definitions

Structs

ArrayString

Represents a C-like string of a given size (mostly like a C char[S] array)

ArrayWideString

Represents a C-like 16-bit string of a given size (mostly like a C char16_t[S] array)

ConstPointerAndSize

Represents a pair of a pointer and a size

PointerAndSize

Represents a pair of a pointer and a size

Uuid

Represents a 16-byte UUID

Functions

simple_panic_handler

Simplified panic handler using a provided Logger type, available as a helpful default panic handler

str_ptr_len⚠

Same as C’s strlen()